+++ paper/paper.bib
@@ -25,13 +25,16 @@
   pages = {240–253}
 }
 
-@Article{jureca,
-	Title   = {{JURECA: General-purpose supercomputer at J\"{u}lich Supercomputing Centre}},
-	Author  = {{J\"{u}lich Supercomputing Centre}},
-	Journal = {Journal of large-scale research facilities},
-	Year    = {2016},
-	Volume  = {2, A62},
-  url     = {https://www.fz-juelich.de/ias/jsc/EN/Expertise/Supercomputers/JURECA/JURECA_node.html}
+@article{Krause2016,
+  doi = {10.17815/jlsrf-2-121},
+  url = {https://doi.org/10.17815/jlsrf-2-121},
+  year = {2016},
+  month = mar,
+  publisher = {Forschungszentrum Julich,  Zentralbibliothek},
+  volume = {2},
+  author = {Dorian Krause and Philipp Th\"{o}rnig},
+  title = {{JURECA}: General-purpose supercomputer at J\"{u}lich Supercomputing Centre},
+  journal = {Journal of large-scale research facilities {JLSRF}}
 }
 
 @article{Schmitz2018,
