+++ paper/paper.bib
@@ -25,6 +25,19 @@
   pages = {240–253}
 }
 
+@article{Menzel2020,
+  doi = {10.1103/physrevx.10.021002},
+  url = {https://doi.org/10.1103/physrevx.10.021002},
+  year = {2020},
+  month = apr,
+  publisher = {American Physical Society ({APS})},
+  volume = {10},
+  number = {2},
+  author = {Miriam Menzel and Markus Axer and Hans De Raedt and Irene Costantini and Ludovico Silvestri and Francesco S. Pavone and Katrin Amunts and Kristel Michielsen},
+  title = {Toward a High-Resolution Reconstruction of 3D Nerve Fiber Architectures and Crossings in the Brain Using Light Scattering Measurements and Finite-Difference Time-Domain Simulations},
+  journal = {Physical Review X}
+}
+
 @article{Krause2016,
   doi = {10.17815/jlsrf-2-121},
   url = {https://doi.org/10.17815/jlsrf-2-121},
