### 2021-01-06 #2923 @fmatuschke
Hi @oliviaguest, I'm so sorry I missed this. Apparently the holidays were necessary.
I looked through the document and found the following potential reviewers: vigji(luigi.petrucco@gmail.com), caldarolamartin(m.caldarola@tudelft.nl), spinicist(tobias.wood@kcl.ac.uk), AmirHoseinHadian(amir.h.hadian@gmail.com) and neurofractal(rob.seymour@ucl.ac.uk)
