### 2021-01-19 #2923 @vigji
Hi @fmatuschke! I have an important deadline in 10 days time and I will be able to look carefully at the code only after that has been met.
Moreover I have to warn you in advance that looking very broadly at the package I got the impression it has been developed with Linux in mind (just following the installation instructions gave my macOS compiler some trouble). I can still test it on a Windows, but I would not set up an Ubuntu machine just for the review. If you think you would prefer not to support other OSs because the code would generally run on clusters that use Linux that is obviously fine (it should be specified in the readme), but I could not take up the review. If you're ok with supporting me trying it out on other machines, happy to give it a try, and I will continue the conversation in issues!
