### 2021-01-19 d05f42b
ADD operating system note
