### 2021-01-19 #2923 @fmatuschke
Hi @vigji. Thanks for pointing out your deadline. That is absolutely not a problem. After all, I know the daily work of a scientist all too well.
The current version is indeed for Linux only. I have added a note in the README. For your information, there are plans to support macOS as well. However, due to the current situation, I don't have access to a macOS system to test everything. On Windows, I test it regularly within the WSL. If you are willing to test it, I would be happy about it. But I can certainly understand the inconvenience.
As a self-taught programmer, I have a lot to learn in this regard as well.
