### 2021-01-19 #2923 @vigji
Ok, if you regularly work within WSL I can give that a try! And keep testing things on macOS, if that can help. Will come back to this after the 29th.
