### 2021-01-19 5270ad1
ADD resample output
