### 2021-01-30 3d-pli/fastpli#1 @vigji
Continuous integration
Hi there,
I will start addressing the points that I would like to raise for the JOSS review through issues.

I would start from continuous integration; I found the gitlab-ci.yml file, but I don't see any GitHub workflow in place to ensure test running on at least the master branch. Am I missing something? Maybe I am just not familiar with your CI workflow, but ideally there should be a badge in the README on the testing state of the package!
