### 2021-01-30 3d-pli/fastpli#2 @vigji
Linting
Just skimming through the project I could find several places in which the usage of linting tools such as `flake8` and `black` would help, eg:
- inconsistent single and double quote  usage;
- unused import statements
- unused, commented out code (this should not live in master)

Ideally those can be running together with the CI mentioned in #1 for testing linting together with test passing.
