### 2021-01-30 3d-pli/fastpli#3 @vigji
Add dependencies versions
I am trying to get the testing running on the WSL. I run into 2 problems so far
- The dependencies do not specify minimum version requirements. Some tests were failing for me as `np.cross` requires `numpy>=1.19`
- Running the tests a second times was failing at a very preliminary stage as a temporary file `/tmp/simply_rep.h5` could not be written as it was already present. Maybe there should be a more gracious handling of temporary file removal if tests fail
