### 2021-01-30 3d-pli/fastpli#4 @vigji
Hosting docs online
It would be highly recommended if the full documentation of the package is hosted somewhere, on GitHub Pages or ReadTheDocs, instead of having this to be run on the user's side and stored locally (in this case, the `sphinx-rtd-theme` and the `sphinx-autodoc-typehints` deps are missing)
