### 2021-01-30 3d-pli/fastpli#4 @vigji
Also, once fixed the dependency issue the docs index.html file is an empty page
