### 2021-01-30 3d-pli/fastpli#5 @vigji
Clarification on the representation of a fiber
I would like to as for clarification as to why a Fiber is not described in Python as its own class, it would probably be better to validate the input data for generating it and ensuring without further checking that all subsequent functions operating on Fibers, FiberBundles etc get consistent input.

Regardless, I was also a bit confused at first that in the module named objects all the transformations are functions, but capitalised as classes should; is there a reason? I have to say that I am still trying to fully understand the interaction between the C++-compiled part of the code, and the Python parts that make use of such modules; if it is related to that, I would be happy to be given additional pointers on the topic. Probably it would be good if a documentation section could also deal with explaining a bit further the C++/Python integration. If there is such file and I missed it apologies, and please link it to me!

Actually if it is in the user generated documentation, see #4
