### 2021-01-30 3d-pli/fastpli#6 @vigji
Expand wiki documentation
I have been reading through the wiki and I can give some feedback on unclear points/things that I think might help clarifying stuff for the user. I apologise in advance if I am saying stuff just because I missed pieces, I just want to give fully constructive feedback from a completely external eye! :)

`fastPLI is an open source toolbox for modeling nerve fibers, simulating them in a 3D PLI [I would probably have it as Polarized light imaging at least in the package line description] microscope and --the-- [simulation?] signal processing.`

I take it is not the case, but I would make it even clearer if it is just for simulation, or also for PLI data analysis (`fastpli.analysis` is the first module you read about in the README, and you might wonder if you can use the package also to just analyse real world data you might provide).

Probably it would be good to guide the user through a reasonable reading order, either with a sequential list of readings in the home of the user that a person should go through, or a reasonable order of pages on the right. Eg, you might want to have them reading about how FiberBundles are represented as the first step. Moreover, you probably want in the organisation of such reading to split even more clearly everything that is related to the generation of the bundles (the sandbox, collision solving, etc); from the reading you would need to do if you just want to simulate the PLI results from the fibres.

I think that adding some intro lines to every page where you showcase code might be useful (as you do eg in https://github.com/3d-pli/fastpli/wiki/Simulation but not in https://github.com/3d-pli/fastpli/wiki/Crossing).

Finally, for users to start get hands in in the package usage, it would be great if those wiki pages could also be distributed as tutorial jupyter notebooks in the package!
