### 2021-01-30 3d-pli/fastpli#5 @vigji
Related to this, reading in the tutorials the representation of FiberBundles properties in a separate list from the list of FiberBundles feels a bit funny (when you define `simpli.fiber_bundles_properties = [[(1.0, -0.001, 10, 'p')]]`). Is there a reason not to keep in a single object a FiberBundle and its properties?
