### 2021-02-02 #2923 @vigji
For the sake of keeping this conversation updated, I started the review through issues in the considered repo, as to my understanding that is the way of proceeding right? My plan would be to report here after the points there have been addressed
