+++ tests/model/sandbox/sandbox_test.py
@@ -6,7 +6,6 @@ import fastpli.objects as obj
 
 
 class MainTest(unittest.TestCase):
-
     def setUp(self):
         pass
 
