+++ src/fastpli/simulation/simpli_cpu.py
@@ -20,7 +20,6 @@ class Simpli(__Simpli):
     """
     Simpli Class for simulating 3D-PLI images
     """
-
     def __init__(self, mpi_comm=None):
 
         super().__init__()
@@ -46,7 +45,8 @@ class Simpli(__Simpli):
     @interpolate.setter
     def interpolate(self, interpolate):
         if interpolate != "NN" and interpolate != "Lerp" and interpolate != "Slerp":
-            raise ValueError("Only \"NN\", \"Lerp\" or \"Slerp\" are supported")
+            raise ValueError(
+                "Only \"NN\", \"Lerp\" or \"Slerp\" are supported")
         self._interpolate = interpolate
 
     def generate_tissue(self, only_tissue=False):
@@ -243,8 +243,8 @@ class Simpli(__Simpli):
 
         flag_rofl = True
         if np.any(self._tilts[:, 1] != np.deg2rad([0, 0, 90, 180, 270])
-                 ) or self._tilts[0, 0] != 0 or np.any(
-                     self._tilts[1:, 0] != self._tilts[1, 0]):
+                  ) or self._tilts[0, 0] != 0 or np.any(
+                      self._tilts[1:, 0] != self._tilts[1, 0]):
             warnings.warn("Tilts not suitable for ROFL. Skipping analysis")
             flag_rofl = False
 
@@ -314,10 +314,10 @@ class Simpli(__Simpli):
 
         if flag_rofl:
             rofl_direction, rofl_incl, rofl_t_rel, (
-                rofl_direction_conf, rofl_incl_conf, rofl_t_rel_conf, rofl_func,
-                rofl_n_iter) = self.apply_rofl(tilting_stack,
-                                               mask=None,
-                                               mp_pool=mp_pool)
+                rofl_direction_conf, rofl_incl_conf, rofl_t_rel_conf,
+                rofl_func, rofl_n_iter) = self.apply_rofl(tilting_stack,
+                                                          mask=None,
+                                                          mp_pool=mp_pool)
         else:
             rofl_direction = None
             rofl_incl = None
@@ -363,7 +363,8 @@ class Simpli(__Simpli):
         self._print("Run pipeline")
         if 'all' in save:
             save = [
-                'tissue', 'optical_axis', 'data', 'optic', 'epa', 'mask', 'rofl'
+                'tissue', 'optical_axis', 'data', 'optic', 'epa', 'mask',
+                'rofl'
             ]
         if 'tissue' in save:
             save = save + ['tissue', 'optical_axis']
