### 2021-02-03 e24db4e
RFC flake8 pep8 conform
