+++ src/fastpli/analysis/rofl.py
@@ -22,7 +22,8 @@ def rofl(data,
     Parameters
     ----------
     data : (tilts, rho)-array_like
-        tilts must be [(0,0), (tilt_angle,0), (tilt_angle,pi/2), (tilt_angle,pi), (tilt_angle,3/2*pi)]
+        tilts must be [(0,0), (tilt_angle,0), (tilt_angle,pi/2),
+                       (tilt_angle,pi), (tilt_angle,3/2*pi)]
         rho must be equidistance between [0,pi)
 
     Returns
