+++ src/fastpli/__init__.py
@@ -1,7 +1,0 @@
-from .__version import __version__
-from .__version import __git__name__
-from .__version import __git__branch__
-from .__version import __git__hash__
-from .__version import __git__build__
-from .__version import __compiler__
-from .__version import __libraries__
