+++ src/fastpli/objects/__init__.py
@@ -1,3 +1,9 @@
 from . import fiber
 from . import fiber_bundle
 from . import fiber_bundles
+
+__all__ = [
+    'fiber',
+    'fiber_bundle',
+    'fiber_bundles',
+]
