+++ src/fastpli/model/solver/__init__.py
@@ -1,1 +1,5 @@
 from .solver import Solver
+
+__all__ = [
+    'Solver',
+]
