+++ src/fastpli/simulation/simpli_cpu.py
@@ -11,7 +11,6 @@ from .. import analysis
 
 import numpy as np
 import warnings
-import sys
 
 from .__simpli import __Simpli
 
@@ -44,14 +43,14 @@ class Simpli(__Simpli):
 
     @interpolate.setter
     def interpolate(self, interpolate):
-        if interpolate != "NN" and interpolate != "Lerp" and interpolate != "Slerp":
+        if interpolate != "NN" and interpolate != "Lerp" and \
+           interpolate != "Slerp":
             raise ValueError(
                 "Only \"NN\", \"Lerp\" or \"Slerp\" are supported")
         self._interpolate = interpolate
 
     def generate_tissue(self, only_tissue=False):
-        """ generating discret tissue for simulation
-        """
+        """ generating discret tissue for simulation """
 
         self._print("Generate Tissue")
         self._print(f"Memory needed: ~{np.ceil(self.memory_usage()):.0f} MB")
@@ -96,7 +95,8 @@ class Simpli(__Simpli):
 
     def run_simulation(self, tissue, optical_axis, tissue_properties, theta,
                        phi):
-        """ running simulation. Input from tissue_generation
+        """
+        running simulation. Input from tissue_generation
 
         tissue and optical_axis will be passed by reference
         theta, phi: tilting angle in radiant
@@ -230,7 +230,9 @@ class Simpli(__Simpli):
                                 save=['data', 'optic', 'epa', 'mask', 'rofl'],
                                 crop_tilt=False,
                                 mp_pool=None):
-        """ Automatic pipeline for simulation and analysis with save options """
+        """
+        Automatic pipeline for simulation and analysis with save options
+        """
 
         self._print("Run simulation pipeline")
         if 'all' in save or 'simulation' in save:
@@ -254,8 +256,8 @@ class Simpli(__Simpli):
         for t, tilt in enumerate(self._tilts):
             theta, phi = tilt[0], tilt[1]
             self._print(
-                f"Tilt {t}: theta: {np.rad2deg(theta):.1f} deg, phi: {np.rad2deg(phi):.1f} deg"
-            )
+                f"Tilt {t}: theta: {np.rad2deg(theta):.1f} deg, phi: " +
+                f"{np.rad2deg(phi):.1f} deg")
             images = self.run_simulation(tissue, optical_axis,
                                          tissue_properties, theta, phi)
 
@@ -358,7 +360,9 @@ class Simpli(__Simpli):
                      ],
                      crop_tilt=False,
                      mp_pool=None):
-        """ Automatic tissue generation and simulation pipeline with save options """
+        """
+        Automatic tissue generation and simulation pipeline with save options
+        """
 
         self._print("Run pipeline")
         if 'all' in save:
