+++ src/fastpli/__init__.py
@@ -1,0 +1,17 @@
+from .__version import __version__
+from .__version import __git__name__
+from .__version import __git__branch__
+from .__version import __git__hash__
+from .__version import __git__build__
+from .__version import __compiler__
+from .__version import __libraries__
+
+__all__ = [
+    '__version__',
+    '__git__name__',
+    '__git__branch__',
+    '__git__hash__',
+    '__git__build__',
+    '__compiler__',
+    '__libraries__',
+]
