### 2021-02-03 648027f
RFC examples flake8 conform
