+++ cmake/setup.py
@@ -15,7 +15,7 @@ setuptools.setup(
     url='http://www.fz-juelich.de/inm/inm-1/EN/Forschung/\
         Fibre%20Architecture/Fibre%20Architecture_node.html',
     python_requires='>3.6.0',
-    install_requires=['numpy', 'numba', 'scipy', 'h5py'],
+    install_requires=['numpy>=1.19', 'numba>=0.52', 'scipy>=1.6', 'h5py>=3.1'],
     zip_safe=False,
     packages=setuptools.find_packages('src'),
     package_dir={'': 'src'},
