+++ examples/crossing.py
@@ -64,7 +64,7 @@ for i in range(1000):
 
     if i == 20:
         solver.draw_scene()
-        solver.fiber_bundles = fastpli.objects.fiber_bundles.ApplyFunToPosition(  # noqa
+        solver.fiber_bundles = fastpli.objects.fiber_bundles.ApplyFunToPosition(  # noqa: E501
             solver.fiber_bundles,
             lambda p: p + np.random.uniform(-10, 10, p.shape))
         solver.fiber_bundles = fastpli.objects.fiber_bundles.ApplyFunToRadii(
