+++ tests/model/solver/solver_test.py
@@ -2,11 +2,10 @@ import unittest
 import numpy as np
 import h5py
 import os
-import tempfile
 
 import fastpli.model.solver
 
-TMP_FILE = os.path.join(tempfile.gettempdir(), "fastpli.test")
+TMP_FILE = os.path.join(os.path.dirname(__file__), "tmp.fastpli.test")
 
 
 class MainTest(unittest.TestCase):
