+++ tests/input_output/fiber_test.py
@@ -2,12 +2,11 @@ import unittest
 import numpy as np
 import h5py
 import os
-import tempfile
 
 from fastpli.io import fiber_bundles
 from fastpli.model.solver import Solver
 
-TMP_FILE = os.path.join(tempfile.gettempdir(), "fastpli.test")
+TMP_FILE = os.path.join(os.path.dirname(__file__), "tmp.fastpli.test.")
 
 
 class MainTest(unittest.TestCase):
@@ -22,26 +21,26 @@ class MainTest(unittest.TestCase):
         self.solver.fiber_bundles = self.fiber_bundles
 
     def test_h5(self):
-        with h5py.File(TMP_FILE + '.h5', 'w-') as h5f:
+        with h5py.File(TMP_FILE + 'fiber_test.h5', 'w') as h5f:
             fiber_bundles.save_h5(h5f, self.solver.fiber_bundles)
 
-        with h5py.File(TMP_FILE + '.h5', 'r') as h5f:
+        with h5py.File(TMP_FILE + 'fiber_test.h5', 'r') as h5f:
             fbs = fiber_bundles.load_h5(h5f)
             for fb_a, fb_b in zip(fbs, self.solver.fiber_bundles):
                 for f_a, f_b in zip(fb_a, fb_b):
                     self.assertTrue(np.alltrue(f_a == f_b))
 
-        with h5py.File(TMP_FILE + '.h5', 'w') as h5f:
+        with h5py.File(TMP_FILE + 'fiber_test.h5', 'w') as h5f:
             h5g = h5f.create_group('test')
             fiber_bundles.save_h5(h5g, self.solver.fiber_bundles)
 
-        with h5py.File(TMP_FILE + '.h5', 'r') as h5f:
+        with h5py.File(TMP_FILE + 'fiber_test.h5', 'r') as h5f:
             fbs = fiber_bundles.load_h5(h5f['test'])
             for fb_a, fb_b in zip(fbs, self.solver.fiber_bundles):
                 for f_a, f_b in zip(fb_a, fb_b):
                     self.assertTrue(np.alltrue(f_a == f_b))
 
-        self.addCleanup(os.remove, TMP_FILE + '.h5')
+        self.addCleanup(os.remove, TMP_FILE + 'fiber_test.h5')
 
     def test_dat(self):
         with open(TMP_FILE + '.dat', 'w') as file:
@@ -56,17 +55,17 @@ class MainTest(unittest.TestCase):
         self.addCleanup(os.remove, TMP_FILE + '.dat')
 
     def test_h5_dat(self):
-        with h5py.File(TMP_FILE + '.h5', 'w-') as h5f:
+        with h5py.File(TMP_FILE + 'fiber_test.h5', 'w') as h5f:
             s = Solver()
             fiber_bundles.save_h5(h5f, s.fiber_bundles)
 
-        with h5py.File(TMP_FILE + '.h5', 'w') as h5f:
+        with h5py.File(TMP_FILE + 'fiber_test.h5', 'w') as h5f:
             fiber_bundles.save_h5(h5f, self.solver.fiber_bundles)
 
         with open(TMP_FILE + '.dat', 'w') as file:
             fiber_bundles.save_dat(file, self.solver.fiber_bundles)
 
-        with h5py.File(TMP_FILE + '.h5', 'r') as h5f:
+        with h5py.File(TMP_FILE + 'fiber_test.h5', 'r') as h5f:
             with open(TMP_FILE + '.dat', 'r') as file:
                 fbs_h5 = fiber_bundles.load_h5(h5f)
                 fbs_dat = fiber_bundles.load_dat(file)
@@ -74,17 +73,18 @@ class MainTest(unittest.TestCase):
[…]
