### 2021-02-03 4a247b2
RFC lower case function names
