+++ tests/objects/fiber_manipulation_test.py
@@ -14,63 +14,63 @@ class MainTest(unittest.TestCase):
         self.fiber_bundles = [[self.fiber.copy()]]
 
     def test_resize(self):
-        fiber = fastpli.objects.fiber.Rescale(self.fiber, 10)
+        fiber = fastpli.objects.fiber.rescale(self.fiber, 10)
         self.assertTrue(np.array_equal(fiber, self.fiber * 10))
 
-        fb = fastpli.objects.fiber_bundle.Rescale(self.fiber_bundle, 10)
+        fb = fastpli.objects.fiber_bundle.rescale(self.fiber_bundle, 10)
         for f in fb:
             self.assertTrue(np.array_equal(f, self.fiber * 10))
 
-        fbs = fastpli.objects.fiber_bundles.Rescale(self.fiber_bundles, 10)
+        fbs = fastpli.objects.fiber_bundles.rescale(self.fiber_bundles, 10)
         for fb in fbs:
             for f in fb:
                 self.assertTrue(np.array_equal(f, self.fiber * 10))
 
-        fiber = fastpli.objects.fiber.Rescale(self.fiber, 10, mod='points')
+        fiber = fastpli.objects.fiber.rescale(self.fiber, 10, mod='points')
         self.assertTrue(np.array_equal(fiber[:, :-2], self.fiber[:, :-2] * 10))
         self.assertTrue(np.array_equal(fiber[:, -1], self.fiber[:, -1]))
 
-        fiber = fastpli.objects.fiber.Rescale(self.fiber, 10, mod='radii')
+        fiber = fastpli.objects.fiber.rescale(self.fiber, 10, mod='radii')
         self.assertTrue(np.array_equal(fiber[:, :-2], self.fiber[:, :-2]))
         self.assertTrue(np.array_equal(fiber[:, -1], self.fiber[:, -1] * 10))
 
     def test_rotation(self):
-        fiber = fastpli.objects.fiber.Rotate(self.fiber,
+        fiber = fastpli.objects.fiber.rotate(self.fiber,
                                              fastpli.tools.rotation.x(0))
         self.assertTrue(np.array_equal(self.fiber, fiber))
 
-        fiber = fastpli.objects.fiber.Rotate(
+        fiber = fastpli.objects.fiber.rotate(
             self.fiber, fastpli.tools.rotation.x(np.deg2rad(90)))
         self.assertTrue(
             np.allclose(fiber, np.array([[0, 0, 0, 1], [1, -1, 1, 2]])))
 
-        fiber = fastpli.objects.fiber.Rotate(
+        fiber = fastpli.objects.fiber.rotate(
             self.fiber, fastpli.tools.rotation.x(np.deg2rad(90)), [1, 1, 1])
         self.assertTrue(
             np.allclose(fiber, np.array([[0, 2, 0, 1], [1, 1, 1, 2]])))
 
         for f in self.fiber_bundle:
-            fiber = fastpli.objects.fiber.Rotate(
+            fiber = fastpli.objects.fiber.rotate(
                 f, fastpli.tools.rotation.x(np.deg2rad(90)), [1, 1, 1])
             self.assertTrue(
                 np.allclose(fiber, np.array([[0, 2, 0, 1], [1, 1, 1, 2]])))
 
         for fb in self.fiber_bundles:
             for f in fb:
-                fiber = fastpli.objects.fiber.Rotate(
+                fiber = fastpli.objects.fiber.rotate(
                     f, fastpli.tools.rotation.x(np.deg2rad(90)), [1, 1, 1])
                 self.assertTrue(
[…]
