+++ src/fastpli/model/solver/solver.py
@@ -20,6 +20,16 @@ import os
 class Solver(_Solver):
     """
     Solver Class for solving collisions between fibers
+
+    Attributes
+    ----------
+    num_steps : int
+       Number of steps.
+
+    Methods
+    -------
+    apply_boundary_conditions(n_max = 10):
+        applies the defined boundry conditions to the inserted model
     """
 
     __is_frozen = False
@@ -152,7 +162,9 @@ class Solver(_Solver):
         """
         Applies collision solving algorithm for one step
 
-        Returns True if solved
+        Returns
+        -------
+        bool: status if the model is solved
         """
         self._num_steps += 1
 
@@ -183,6 +195,14 @@ class Solver(_Solver):
     def apply_boundary_conditions(self, n_max=10):
         """
         Applies boundary conditions for n_max steps without collision solving.
+
+        Parameters
+        ----------
+        n_max: (optional) maximal number of iterations
+
+        Returns
+        -------
+        [[(nx4)-arrays]]: fiber bundles
         """
         if not isinstance(n_max, int) or n_max <= 0:
             raise TypeError("only integer > 0 allowed")
