+++ .github/workflows/main.yml
@@ -1,0 +1,36 @@
+
+name: CI
+
+# Controls when the action will run. 
+on:
+  # Triggers the workflow on push or pull request events but only for the master branch
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+jobs:
+  test:
+    runs-on: ubuntu:20.04
+
+    steps:
+      - dependencies:
+        run: |
+          export DEBIAN_FRONTEND=noninteractive
+          apt-get update -qq && apt-get upgrade -y
+          apt-get install -y gcc-9 g++-9 clang-10 clang-format-10 cmake make git
+          apt-get install -y python3-dev python3-venv
+          apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
+          apt-get install -y freeglut3-dev
+
+      - testing:
+        run: |
+          docker/.docker_run.sh
+          apt-get remove -y freeglut3-dev
+          apt-get autoremove -y
+          make clean
+          make BUILD=debug CC=gcc-9 CXX=g++-9 install
+          make test
