+++ examples/crossing.py
@@ -67,7 +67,7 @@ for i in range(1000):
         solver.fiber_bundles = fastpli.objects.fiber_bundles.apply_fun_to_position(  # noqa: E501
             solver.fiber_bundles,
             lambda p: p + np.random.uniform(-10, 10, p.shape))
-        solver.fiber_bundles = fastpli.objects.fiber_bundles.apply_fun_to_radii(
+        solver.fiber_bundles = fastpli.objects.fiber_bundles.apply_fun_to_radii(  # noqa: E501
             solver.fiber_bundles,
             lambda r: r * np.random.lognormal(0, 0.1, r.shape))
         solver.draw_scene()
