+++ .github/workflows/main.yml
@@ -17,7 +17,7 @@ jobs:
     runs-on: ubuntu:20.04
 
     steps:
-      - dependencies:
+      - name: dependencies
         run: |
           export DEBIAN_FRONTEND=noninteractive
           apt-get update -qq && apt-get upgrade -y
@@ -26,7 +26,7 @@ jobs:
           apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
           apt-get install -y freeglut3-dev
 
-      - testing:
+      - name: testing
         run: |
           docker/.docker_run.sh
           apt-get remove -y freeglut3-dev
