+++ .github/workflows/main.yml
@@ -14,7 +14,7 @@ on:
 
 jobs:
   test:
-    runs-on: ubuntu:20.04
+    runs-on: ubuntu-20.04
 
     steps:
       - name: dependencies
