+++ .github/workflows/main.yml
@@ -13,24 +13,25 @@ on:
   workflow_dispatch:
 
 jobs:
-  test:
+  Testing:
     runs-on: ubuntu-20.04
 
     steps:
       - name: dependencies
         run: |
           export DEBIAN_FRONTEND=noninteractive
-          apt-get update -qq && apt-get upgrade -y
-          apt-get install -y gcc-9 g++-9 clang-10 clang-format-10 cmake make git
-          apt-get install -y python3-dev python3-venv
-          apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
-          apt-get install -y freeglut3-dev
+          sudo apt-get update -qq
+          sudo apt-get upgrade -y
+          sudo apt-get install -y gcc-9 g++-9 clang-10 clang-format-10 cmake make git
+          sudo apt-get install -y python3-dev python3-venv
+          sudo apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
+          sudo apt-get install -y freeglut3-dev
 
       - name: testing
         run: |
           docker/.docker_run.sh
-          apt-get remove -y freeglut3-dev
-          apt-get autoremove -y
+          sudo apt-get remove -y freeglut3-dev
+          sudo apt-get autoremove -y
           make clean
           make BUILD=debug CC=gcc-9 CXX=g++-9 install
           make test
