+++ .github/workflows/main.yml
@@ -21,7 +21,6 @@ jobs:
         run: |
           export DEBIAN_FRONTEND=noninteractive
           sudo apt-get update -qq
-          sudo apt-get upgrade -y
           sudo apt-get install -y gcc-9 g++-9 clang-10 clang-format-10 cmake make git
           sudo apt-get install -y python3-dev python3-venv
           sudo apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
