### 2021-02-03 eee0d8a
FIX version dependencies for python3.6
