+++ .github/workflows/main.yml
@@ -5,9 +5,9 @@ name: CI
 on:
   # Triggers the workflow on push or pull request events but only for the master branch
   push:
-    branches: [ master, pull_request ]
+    branches: [ master, TestCI ]
   pull_request:
-    branches: [ master, pull_request ]
+    branches: [ master ]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
@@ -28,7 +28,6 @@ jobs:
           sudo apt-get install -y python3-dev python3-venv
           sudo apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
           sudo apt-get install -y freeglut3-dev
-          cd $GITHUB_WORKSPACE
 
       - name: run
         run: |
