+++ .github/workflows/main.yml
@@ -16,18 +16,24 @@ jobs:
   Testing:
     runs-on: ubuntu-latest
 
+    strategy:
+      matrix:
+        python-version: [3.6, 3.7, 3.8]
+
     steps:
       - uses: actions/checkout@v2
-        with:
-          fetch-depth: 1
+
       - name: dependencies
         run: |
           export DEBIAN_FRONTEND=noninteractive
           sudo apt-get update -qq
-          sudo apt-get install -y gcc-9 g++-9 clang-10 clang-format-10 cmake make git
-          sudo apt-get install -y python3-dev python3-venv
+          sudo apt-get install -y gcc-9 g++-9 cmake make git
           sudo apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
           sudo apt-get install -y freeglut3-dev
+          
+      - name: python ${{ matrix.python-version }}
+        run: |
+          sudo apt-get install -y python${{ matrix.python-version }}-dev python${{ matrix.python-version }}-venv
 
       - name: run
         run: |
