### 2021-02-03 8cba972
RFC CI, adjusted compiler and python version to OS version
