+++ .gitlab/.gitlab-ci.yml
@@ -3,17 +3,6 @@ image: ubuntu:20.04
 stages:
    - test
 
-before_script:
-   - export DEBIAN_FRONTEND=noninteractive
-   - apt-get update -qq && apt-get upgrade -y
-   - apt-get install -y gcc-9 g++-9 clang-10 clang-format-10 cmake make git
-   - apt-get install -y python3-dev python3-venv
-   - apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
-   - apt-get install -y freeglut3-dev
-
-# TODO: test-mpi 
-# - export HDF5_DIR="/usr/lib/x86_64-linux-gnu/hdf5/openmpi"
-
 test:release:
    stage: test
    only: 
@@ -26,9 +15,29 @@ test:release:
       variables:
          - $CI_COMMIT_MESSAGE =~ /WIP/
    script:
+      - export DEBIAN_FRONTEND=noninteractive
+      - apt-get update -qq
+      - apt-get install -y gcc g++ clang-format-10 cmake make git
+      - apt-get install -y python3-dev python3-venv
+      - apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
+      - apt-get install -y freeglut3-dev
       - docker/.docker_run.sh
-      - apt-get remove -y freeglut3-dev
-      - apt-get autoremove -y
-      - make clean
-      - make BUILD=debug CC=gcc-9 CXX=g++-9 install
-      - make test
+
+test:no_opengl:
+   stage: test
+   only: 
+      - development
+      - master
+      - TestCI
+   tags:
+      - docker
+   except:
+      variables:
+         - $CI_COMMIT_MESSAGE =~ /WIP/
+   script:
+      - export DEBIAN_FRONTEND=noninteractive
+      - apt-get update -qq
+      - apt-get install -y gcc g++ clang-format-10 cmake make git
+      - apt-get install -y python3-dev python3-venv
+      - apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
+      - docker/.docker_run.sh
