+++ CMakeLists.txt
@@ -84,8 +84,8 @@ ADD_SUBDIRECTORY(lib/pybind11)
 IF (CMAKE_CXX_COMPILER_ID STREQUAL "Clang")
    SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fsized-deallocation" )
 ENDIF()
-LIST(APPEND FASTPLI_INFO_LIBRARIES "PythonInterp-v${PythonInterp_VERSION}")
-LIST(APPEND FASTPLI_INFO_LIBRARIES "PythonLibs-v${PythonLibs_VERSION}")
+INCLUDE_DIRECTORIES(${Python3_INCLUDE_DIRS})
+LIST(APPEND FASTPLI_INFO_LIBRARIES "Python-v${Python3_VERSION}")
 
 # OpenMP
 FIND_PACKAGE( OpenMP )
