+++ .github/workflows/main.yml
@@ -5,7 +5,7 @@ name: CI
 on:
   # Triggers the workflow on push or pull request events but only for the master branch
   push:
-    branches: [ master, TestCI ]
+    branches: [ master, development, TestCI ]
   pull_request:
     branches: [ master ]
 
