### 2021-02-03 2130a4f
[!!!] RM actions/python because of building Issue. ADD ubuntu18.04 for testing python 3.6.
