+++ .github/workflows/main.yml
@@ -14,11 +14,11 @@ on:
 
 jobs:
   Testing:
-    runs-on: ubuntu-20.04
-
     strategy:
       matrix:
-        python-version: [3.6, 3.7, 3.8]
+        os: [ubuntu-18.04, ubuntu-20.04]
+
+    runs-on: ${{ matrix.os }}
 
     steps:
       - uses: actions/checkout@v2
@@ -28,13 +28,9 @@ jobs:
           export DEBIAN_FRONTEND=noninteractive
           sudo apt-get update -qq
           sudo apt-get install -y gcc g++ clang-format-10 cmake make git
+          sudo apt-get install -y python3-dev python3-venv
           sudo apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
           sudo apt-get install -y freeglut3-dev
-
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
-        with:
-          python-version: ${{ matrix.python-version }}
           
       - name: run
         run: |
