+++ src/model/solver/world.cpp
@@ -8,6 +8,8 @@
 #include <memory>
 #include <utility>
 
+#include <Python.h>
+
 #include "fiber_class.hpp"
 #include "include/aabb.hpp"
 #include "include/omp.hpp"
