+++ .github/workflows/main.yml
@@ -17,6 +17,9 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-18.04, ubuntu-20.04]
+        config:
+        - {run-opengl: true}
+        - {run-opengl: false}
 
     runs-on: ${{ matrix.os }}
 
@@ -30,7 +33,11 @@ jobs:
           sudo apt-get install -y gcc g++ clang-format-10 cmake make git
           sudo apt-get install -y python3-dev python3-venv
           sudo apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
+
+      - name: openGL
+        run:
           sudo apt-get install -y freeglut3-dev
+        if: matrix.config.run-it
           
       - name: run
         run: |
