### 2021-02-04 e79a783
WIP CI scripts
