+++ CI/run-env.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -euo pipefail
+
+python3 -m venv env-CI
