+++ CI/dockerfile
@@ -1,9 +1,9 @@
-FROM ubuntu:20.04
+FROM ubuntu:latest
 
 ENV DEBIAN_FRONTEND=noninteractive
 
 RUN apt-get update
-RUN apt-get install -y clang clang-format-10 make cmake git
+RUN apt-get install -y clang clang-format-11 make cmake git
 RUN apt-get install -y python3-dev python3-venv
 RUN apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
 RUN apt-get install -y freeglut3-dev
@@ -12,4 +12,4 @@ ENV HDF5_DIR /usr/lib/x86_64-linux-gnu/hdf5/openmpi
 
 WORKDIR /code/fastpli
 
-CMD [ "docker/.docker_run.sh" ]
+CMD [ "CI/run-all.sh" ]
