### 2021-02-04 a5caf54
ADD CI scripts, RFC test structure
