+++ CI/run-build.sh
@@ -1,9 +1,6 @@
 #!/bin/bash
 set -euo pipefail
 
-echo "--------------------------------------"
-echo "****************Build*****************"
-echo "--------------------------------------"
 make fastpli
 
 env-CI/bin/pip3 install --upgrade pip
