+++ .github/workflows/main.yml
@@ -1,15 +1,12 @@
 
 name: CI
 
-# Controls when the action will run. 
 on:
-  # Triggers the workflow on push or pull request events but only for the master branch
   push:
     branches: [ master, development, TestCI ]
   pull_request:
     branches: [ master ]
 
-  # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
 
 jobs:
@@ -41,4 +38,4 @@ jobs:
           
       - name: run
         run: |
-          docker/.docker_run.sh
+          ./CI/run_all.sh
