+++ .github/workflows/main.yml
@@ -38,4 +38,4 @@ jobs:
           
       - name: run
         run: |
-          ./CI/run_all.sh
+          ./CI/run-all.sh
