+++ .gitlab/.gitlab-ci.yml
@@ -21,7 +21,7 @@ test:release:
       - apt-get install -y python3-dev python3-venv
       - apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
       - apt-get install -y freeglut3-dev
-      - ./CI/run_all.sh
+      - CI/run-all.sh
 
 test:no_opengl:
    stage: test
@@ -40,4 +40,4 @@ test:no_opengl:
       - apt-get install -y gcc g++ clang-format-10 cmake make git
       - apt-get install -y python3-dev python3-venv
       - apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
-      - ./CI/run_all.sh
+      - CI/run-all.sh
