+++ .gitlab/.gitlab-ci.yml
@@ -23,21 +23,21 @@ test:release:
       - apt-get install -y freeglut3-dev
       - CI/run-all.sh
 
-test:no_opengl:
-   stage: test
-   only: 
-      - development
-      - master
-      - TestCI
-   tags:
-      - docker
-   except:
-      variables:
-         - $CI_COMMIT_MESSAGE =~ /WIP/
-   script:
-      - export DEBIAN_FRONTEND=noninteractive
-      - apt-get update -qq
-      - apt-get install -y gcc g++ clang-format-10 cmake make git
-      - apt-get install -y python3-dev python3-venv
-      - apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
-      - CI/run-all.sh
+# test:no_opengl:
+#    stage: test
+#    only: 
+#       - development
+#       - master
+#       - TestCI
+#    tags:
+#       - docker
+#    except:
+#       variables:
+#          - $CI_COMMIT_MESSAGE =~ /WIP/
+#    script:
+#       - export DEBIAN_FRONTEND=noninteractive
+#       - apt-get update -qq
+#       - apt-get install -y gcc g++ clang-format-10 cmake make git
+#       - apt-get install -y python3-dev python3-venv
+#       - apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
+#       - CI/run-all.sh
