### 2021-02-05 037cddd
ADD automatic code formatter, RFC code checker, ADD format config files
