+++ tests/fastpli_test.py
@@ -2,6 +2,7 @@ import unittest
 
 
 class MainTest(unittest.TestCase):
+
     def test_import(self):
         import fastpli
         import fastpli.analysis
