+++ src/fastpli/analysis/affine_transformation.py
@@ -49,8 +49,7 @@ def _interpolate_griddata(image, M, mode):
 
     for k in range(image.shape[2]):
         image_nn[:, :, k] = scipy.interpolate.griddata(points.T,
-                                                       image[:, :,
-                                                             k].flatten(),
+                                                       image[:, :, k].flatten(),
                                                        (grid_i, grid_j),
                                                        method=mode)
 
@@ -84,7 +83,8 @@ def calc_matrix(p_in, p_out):
     L = p_in.shape[0]
     B = np.vstack([np.transpose(p_in), np.ones(L)])
     D = 1.0 / np.linalg.det(B)
-    M = np.array([[(-1)**i * D * _replace_mat_row(B, R, i) for i in range(L)]
+    M = np.array([[(-1)**i * D * _replace_mat_row(B, R, i)
+                   for i in range(L)]
                   for R in np.transpose(p_out)])
 
     return np.vstack([M, [0, 0, 1]])
