+++ src/fastpli/objects/fiber_bundles.py
@@ -69,8 +69,7 @@ def rescale(fiber_bundles, scale, mod='all'):
     fiber_bundles = copy.deepcopy(fiber_bundles)
     for j, fb in enumerate(fiber_bundles):
         for i, _ in enumerate(fb):
-            fiber_bundles[j][i] = fiber.rescale(fiber_bundles[j][i], scale,
-                                                mod)
+            fiber_bundles[j][i] = fiber.rescale(fiber_bundles[j][i], scale, mod)
     return fiber_bundles
 
 
@@ -99,8 +98,7 @@ def rotate(fiber_bundles, rot, offset=None):
         offset = np.array(offset, copy=False)
     for j, fb in enumerate(fiber_bundles):
         for i, _ in enumerate(fb):
-            fiber_bundles[j][i] = fiber.rotate(fiber_bundles[j][i], rot,
-                                               offset)
+            fiber_bundles[j][i] = fiber.rotate(fiber_bundles[j][i], rot, offset)
     return fiber_bundles
 
 
