+++ tests/model/solver/solver_test.py
@@ -9,6 +9,7 @@ TMP_FILE = os.path.join(os.path.dirname(__file__), "tmp.fastpli.test")
 
 
 class MainTest(unittest.TestCase):
+
     def setUp(self):
         self._test_fiber = np.array([[0, 0, 0, 1], [0, 0, 1, 2]])
         self._test_fiberbundles = [[self._test_fiber]]

[…]
