+++ tests/test.py
@@ -1,6 +1,6 @@
 import unittest
 
-if __name__ == "__main__":
-    suite = unittest.TestLoader().discover('.', pattern="*_test.py")
+if __name__ == '__main__':
+    suite = unittest.TestLoader().discover('.', pattern='*_test.py')
     result = unittest.TextTestRunner(verbosity=2).run(suite)
     exit(not result.wasSuccessful())
