### 2021-02-05 acc3b79
ADD developer options to makefile
