+++ src/fastpli/simulation/simpli_cpu.py
@@ -345,10 +345,10 @@ class Simpli(__Simpli):
             h5f['analysis/rofl/t_rel'] = rofl_t_rel
 
         if h5f and flag_rofl and 'rofl_conf' in save:
-            h5f['analysis/rofl/direction_conf'] = rofl_direction_conf,
-            h5f['analysis/rofl/inclination_conf'] = rofl_incl_conf,
-            h5f['analysis/rofl/t_rel_conf'] = rofl_t_rel_conf,
-            h5f['analysis/rofl/func'] = rofl_func,
+            h5f['analysis/rofl/direction_conf'] = rofl_direction_conf
+            h5f['analysis/rofl/inclination_conf'] = rofl_incl_conf
+            h5f['analysis/rofl/t_rel_conf'] = rofl_t_rel_conf
+            h5f['analysis/rofl/func'] = rofl_func
             h5f['analysis/rofl/n_iter'] = rofl_n_iter
 
         if flag_rofl:
