### 2021-02-05 6e8d252
RFC linting improvements
