+++ src/fastpli/simulation/optic.py
@@ -1,7 +1,8 @@
-import numpy as np
 import warnings
+
 import scipy.ndimage
 import numba
+import numpy as np
 
 
 def add_noise(image, model, mask=None):
