+++ src/fastpli/analysis/_ROFL_with_jacobi.py
@@ -4,8 +4,8 @@ Author: Daniel Schmitz INM-1, Forschungszentrum Jülich:
 https://doi.org/10.3389/fnana.2018.00075
 """
 
-import numpy as np
 import itertools
+import numpy as np
 from scipy import optimize
 
 import numba
@@ -42,7 +42,7 @@ def _invertmatrix(m):
 
 def _calc_centered_vals(stack_2d, gain):
 
-    num_tilts, num_rots = stack_2d.shape
+    _, num_rots = stack_2d.shape
     centered_stack = np.empty_like(stack_2d)
     trans = np.mean(stack_2d, 1)
     trans_extended = np.tile(trans, (num_rots, 1)).T
@@ -315,17 +315,9 @@ def _calc_Jacobi_for_opt(phi, alpha, t_rel, num_rotations):
     return J
 
 
-def _brute_force_grid(phi,
-                      number_inclination_steps,
-                      number_t_rel_steps,
-                      number_rotations,
-                      number_tilt_steps,
-                      array_of_matrices,
-                      dir_offset,
-                      tau,
-                      measures,
-                      weights,
-                      ret=None):
+def _brute_force_grid(phi, number_inclination_steps, number_t_rel_steps,
+                      number_rotations, number_tilt_steps, array_of_matrices,
+                      dir_offset, tau, measures, weights):
     """
     Returns the minimumm of the fitted intensities on the brute force grid
 
@@ -350,84 +342,41 @@ def _brute_force_grid(phi,
                              number_inclination_steps + 2)[1:-1]
     t_rel_steps = np.linspace(0, 0.9, number_t_rel_steps + 2)[1:-1]
 
-    if (ret is not None and ret < 0.05):
+    # initialize output array
+    number_gridpoints = number_inclination_steps * number_t_rel_steps
 
-        dir_steps = np.linspace(0, np.pi, number_inclination_steps + 2)[1:-1]
+    grid_intensities_output = np.empty((number_gridpoints, number_measurements))
 
-        number_gridpoints = number_inclination_steps**2 * number_t_rel_steps
+    j = 0
 
-        grid_intensities_output = np.empty(
-            (number_gridpoints, number_measurements))
+    # initialize list of all incl/trel combinations
+    gridlist = list(itertools.product(incl_steps, t_rel_steps))
 
-        j = 0
+    # loop over all grid points
+    for (incl, trel) in itertools.product(incl_steps, t_rel_steps):
 
-        # initialize list of all incl/trel combinations
-        gridlist = list(itertools.product(dir_steps, incl_steps, t_rel_steps))
+        # calculate tilted angles and trel
+        phi_array, alpha_array, t_rel_array = _calculate_rotated_fiber_params(
+            phi, incl, trel, array_of_matrices, tau)
 
-        # loop over all grid points
-        for (phi, incl, trel) in itertools.product(dir_steps, incl_steps,
-                                                   t_rel_steps):
+        # evaluate function and write into output array
+        grid_intensities_output[j, :] = _calc_Int_single_fiber_fitted(
+            phi_array, alpha_array, t_rel_array, number_rotations, dir_offset)
+        j += 1
 
-            # calculate tilted angles and trel
-            phi_array, alpha_array, t_rel_array = _calculate_rotated_fiber_params(  # noqa: E501
-                phi, incl, trel, array_of_matrices, tau)
+    # broadcast measurement array number_gridpoints times so it can be
+    # substracted from all grid point evaluations
+    measures_repeat = np.broadcast_to(measures,
+                                      (number_gridpoints, number_measurements))
 
-            # evaluate function and write into outputarray
-            grid_intensities_output[j, :] = _calc_Int_single_fiber_fitted(
[…]
