### 2021-02-05 cc25975
ADD possible speedup comment
