+++ CI/run-build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 set -euo pipefail
 
-make fastpli
+make clean
+make fastpli $1
 
 env-CI/bin/pip3 install --upgrade pip
 env-CI/bin/pip3 install .
