+++ CI/run-all.sh
@@ -8,7 +8,7 @@ echo "--------------------------------------"
 echo "--------------------------------------"
 echo "****************Build*****************"
 echo "--------------------------------------"
-./CI/run-build.sh
+./CI/run-build.sh $1
 echo "--------------------------------------"
 echo "*****************Code*****************"
 echo "--------------------------------------"
