+++ CI/run-build.sh
@@ -2,7 +2,12 @@
 set -euo pipefail
 
 make clean
-make fastpli $1
+
+if [ $# -eq 0 ]; then
+   make fastpli
+else
+   make fastpli $1
+fi
 
 env-CI/bin/pip3 install --upgrade pip
 env-CI/bin/pip3 install .
