+++ CI/run-all.sh
@@ -8,7 +8,11 @@ echo "--------------------------------------"
 echo "--------------------------------------"
 echo "****************Build*****************"
 echo "--------------------------------------"
-./CI/run-build.sh $1
+if [ $# -eq 0 ]; then
+   ./CI/run-build.sh
+else
+   ./CI/run-build.sh $1
+fi
 echo "--------------------------------------"
 echo "*****************Code*****************"
 echo "--------------------------------------"
