### 2021-02-05 06c620b
WIP jupyter notebooks tutorial
