+++ examples/requirements.txt
@@ -1,3 +1,3 @@
+jupyter
 matplotlib
-mpi4py
-imageio
+PyQt5
