+++ CI/run-all.sh
@@ -5,6 +5,7 @@ echo "--------------------------------------"
 echo "****************env-CI****************"
 echo "--------------------------------------"
 ./CI/run-env.sh
+echo "Done"
 echo "--------------------------------------"
 echo "****************Build*****************"
 echo "--------------------------------------"
@@ -13,18 +14,19 @@ if [ $# -eq 0 ]; then
 else
    ./CI/run-build.sh $1
 fi
+echo "Done"
 echo "--------------------------------------"
 echo "*****************Code*****************"
 echo "--------------------------------------"
 ./CI/run-code-check.sh
+echo "Done"
 echo "--------------------------------------"
 echo "***************Examples***************"
 echo "--------------------------------------"
 ./CI/run-examples.sh
+echo "Done"
 echo "--------------------------------------"
 echo "****************Tests*****************"
 echo "--------------------------------------"
 ./CI/run-pytest.sh
-echo "--------------------------------------"
-echo "*****************Done*****************"
-echo "--------------------------------------"
+echo "Done"
