### 2021-02-06 ed0657a
RM imageio from examples, using matplotlib only
