### 2021-02-06 2524937
FIX multiple run of solver example
