+++ CI/run-all.sh
@@ -5,7 +5,7 @@ echo "--------------------------------------"
 echo "****************env-CI****************"
 echo "--------------------------------------"
 ./CI/run-env.sh
-echo "-----------------Done-----------------"
+echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
 echo "--------------------------------------"
 echo "****************Build*****************"
 echo "--------------------------------------"
@@ -14,19 +14,19 @@ if [ $# -eq 0 ]; then
 else
    ./CI/run-build.sh $1
 fi
-echo "-----------------Done-----------------"
+echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
 echo "--------------------------------------"
 echo "*************Code-Format**************"
 echo "--------------------------------------"
 ./CI/run-code-check.sh
-echo "-----------------Done-----------------"
+echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
 echo "--------------------------------------"
 echo "****************Tests*****************"
 echo "--------------------------------------"
 ./CI/run-pytest.sh
-echo "-----------------Done-----------------"
+echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
 echo "--------------------------------------"
 echo "***************Examples***************"
 echo "--------------------------------------"
 ./CI/run-examples.sh
-echo "-----------------Done-----------------"
+echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
