### 2021-02-06 abeec6e
RFC variable naming
