+++ _sources/_autosummary/fastpli.analysis.rofl.rofl.rst.txt
@@ -1,0 +1,6 @@
+fastpli.analysis.rofl.rofl
+==========================
+
+.. currentmodule:: fastpli.analysis.rofl
+
+.. autofunction:: rofl
