+++ _sources/_autosummary/fastpli.simulation.optic.filter_resize.rst.txt
@@ -1,0 +1,6 @@
+fastpli.simulation.optic.filter\_resize
+=======================================
+
+.. currentmodule:: fastpli.simulation.optic
+
+.. autofunction:: filter_resize
