+++ _sources/_autosummary/fastpli.tools.rotation.y.rst.txt
@@ -1,0 +1,6 @@
+fastpli.tools.rotation.y
+========================
+
+.. currentmodule:: fastpli.tools.rotation
+
+.. autofunction:: y
