+++ _sources/_autosummary/fastpli.analysis.images.unit_vectors.rst.txt
@@ -1,0 +1,6 @@
+fastpli.analysis.images.unit\_vectors
+=====================================
+
+.. currentmodule:: fastpli.analysis.images
+
+.. autofunction:: unit_vectors
