+++ _sources/_autosummary/fastpli.io.fiber_bundles.load.rst.txt
@@ -1,0 +1,6 @@
+fastpli.io.fiber\_bundles.load
+==============================
+
+.. currentmodule:: fastpli.io.fiber_bundles
+
+.. autofunction:: load
