### 2021-02-06 4394e79
ADD manuel sphinx to gh-pages build
