+++ docs/source/conf.py
@@ -23,7 +23,7 @@ project = 'fastpli'
 author = 'fmatuschke'
 
 # The full version, including alpha/beta/rc tags
-release = '0.9'
+release = '1.0'
 
 # -- General configuration ---------------------------------------------------
 
