+++ .github/workflows/docs.yml
@@ -1,0 +1,32 @@
+name: Documentation
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+  workflow_dispatch:
+
+jobs:
+  Sphinx-Doc:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: dependencies
+        run: |
+          export DEBIAN_FRONTEND=noninteractive
+          sudo apt-get update -qq
+          sudo apt-get install -y make git
+          sudo apt-get install -y python3-dev python3-venv
+
+      - name: build
+        run: |
+          commit_msg=$(git describe --always)
+          make docs
+          cd docs/gh-pages
+          git add .
+          git commit -m "$commit_msg"
+          git push
