### 2021-02-06 b238615
ADD sphinx deployment to main.yml for testing
