### 2021-02-06 f195e86
ADD docs to CI and docker
