+++ CI/run-all.sh
@@ -6,6 +6,7 @@ echo "****************env-CI****************"
 echo "--------------------------------------"
 ./CI/run-env.sh
 echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
+
 echo "--------------------------------------"
 echo "****************Build*****************"
 echo "--------------------------------------"
@@ -15,18 +16,27 @@ else
    ./CI/run-build.sh $1
 fi
 echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
+
 echo "--------------------------------------"
 echo "*************Code-Format**************"
 echo "--------------------------------------"
 ./CI/run-code-check.sh
 echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
+
 echo "--------------------------------------"
 echo "****************Tests*****************"
 echo "--------------------------------------"
 ./CI/run-pytest.sh
 echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
+
 echo "--------------------------------------"
 echo "***************Examples***************"
 echo "--------------------------------------"
 ./CI/run-examples.sh
+echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
+
+echo "--------------------------------------"
+echo "*****************Docs*****************"
+echo "--------------------------------------"
+./CI/run-docs.sh
 echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
