### 2021-02-07 ba71972
WIP restructure documentation
