+++ README.md
@@ -156,4 +156,4 @@ are also welcome to add features or correct problems.
 
 ## License
 
-This project is licensed under the GPLv3 License - see the [LICENSE](LICENSE) file for details
+This project is licensed under the MIT License - see the [LICENSE](LICENSE) file for details
