### 2021-02-08 3d-pli/fastpli#1 @fmatuschke
Hi @vigji,

The gitlab-ci.yml is from our private GitLab institute server. As this project becomes fully publicly available, I am moving everything to GitHub.

I managed to integrate all functionalities to GitHub actions like running tests. I also added code formatter and linting actions (#2) and automatic deployments of a GitHub page with the latest version of a module documentation (#4, still WIP). Additionally, I am currently also working on an automatic release of the software at every release version of this project.

FYI: the current developing process is inside the [development](https://github.com/3d-pli/fastpli/tree/development) branch. Since I'm changing quite a lot at the moment and many changes are interdependent, it's a bit chaotic at the moment. I'm working towards an official release and will keep you informed.
