### 2021-02-08 3d-pli/fastpli#2 @fmatuschke
Thanks for the suggestions. I added an automatic linting process to my workflow and cleaned everything up. As you suggest they are also implemented inside the CI. To configure the style I added an `.style.yapf` and `.flake8` file to the repository.
