### 2021-02-08 3d-pli/fastpli#3 @fmatuschke
I added the current dependencies to the `setup.py` file. I hope this fixes the Issue.
Additionally, I changed the testing process. Every temporary file is now created inside the test path and cleaned up automatically.
When an error occurs which interrupts the cleanup process, the files will be now overwritten.

FYI, I also added a sentence where additional information about the WSL can be found.
