### 2021-02-08 3d-pli/fastpli#4 @fmatuschke
This is a very good recommendation. I added an automatic process to publish the module documentation to a GitHub Page <https://3d-pli.github.io/fastpli/> at every release step.

As you probably can see, the module documentation is currently not finished. I am working on adding doc strings to all methods and adding additional information to the page. I am also thinking about merging the Wiki to the GitHub Page so that everything is combined. If I find a fast and nice looking solution for this I will integrate it.

I will update everything with a new version of the wiki that also takes #5 and #6 into account.
