### 2021-02-08 3d-pli/fastpli#6 @fmatuschke
Thank you for your suggestions. Your external eye is much appreciated. This is after all the idea of a review :).

I am currently redesigning the Wiki pages. This will still take some time. It is a very good idea to guide the user more through the process of the toolbox.

I also like your idea of a jupyter notebook as an interactive tutorial. Currently, I added first drafts for the `sandbox` and `solver` process inside the `./examples` folder. Their output will also be available at the corresponding Wiki pages.

I will inform you as soon as I have an official release.
