### 2021-02-08 3d-pli/fastpli#6 @vigji
I'll put here a cumulative reply:
Great that the suggestions were appreciated, let me know whenever you think the next release incorporating these changes is ready, I am looking forward to have another look! :)
