### 2021-02-08 3d-pli/fastpli#5 @fmatuschke
I thought about representing the fibers as their own classes. I decided against it because I want to give the user as much freedom as possible to interact with the objects during the building process of models. Additionally, I run into some problems along the pybind11 integration process. But this was already quite some time ago.
You are correct in suggesting a documentation part about the C++ and Python interaction. Since I am currently in redesigning the Wiki (regarding #4, #5 and #6) I will add such a section.

The capitalization of the functions was a mistake. I corrected it.

Regarding your question about the `fiber_bundle_properties`. The `simpli` class is split into two processes. The generation and simulation stage. The `run_generation` function generate a discretized 3d representation of the fiber models. Only for the simulation part the properties are necessary. Additionally, an open question is yet which properties should be used for the fibers. Nerve fibers are surrounded by a myelin. The myelin are "lipid layers" wrapped around the axon. How many layers, thickness, different optical properties is not yet fully known. Therefore, the user is capable of changing the properties on the same geometrical model. Additionally, it is possible to run simulations if a discretized volume is provided. That's why I split the properties from the pure geometrical representation. I will try to make this clear inside the wiki and examples.

FYI, this toolbox is capable of building and simulating very large datasets. I tried to make everything as fast as possible. Additionally, I also had a main focus on trying to prevent copping of data. Between Python and C++ I found it quite easy if I am using NumPy arrays. I will try to make this clear inside the Wiki as well.
