### 2021-02-08 20cabc9
Typos, ADD example/*png to gitignore
