### 2021-02-08 9e3faf7
RFC readme structure consistent with wiki
