+++ examples/sandbox.py
@@ -1,7 +1,6 @@
 import fastpli.model.sandbox as sandbox
 
 import numpy as np
-
 import matplotlib.pyplot as plt
 
 
