+++ src/fastpli/objects/__init__.py
@@ -1,9 +1,4 @@
-from . import fiber
-from . import fiber_bundle
-from . import fiber_bundles
+from .fibers import Fiber, FiberBundle, FiberBundles
+from .layers import Layer, Layers
 
-__all__ = [
-    'fiber',
-    'fiber_bundle',
-    'fiber_bundles',
-]
+__all__ = ['Fiber', 'FiberBundle', 'FiberBundles', 'Layer', 'Layers']
