+++ src/fastpli/objects/fiber.py
@@ -1,213 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Methods for manipulation of fiber objects
-"""
-
-import numpy as np
-import numba
-
-
-def scale(fiber, scale, mode='all'):
-    """
-    Rescales fiber
-
-    Parameters
-    ----------
-    fiber : (,4)-array
-        fiber
-    scale : float
-        scale factor
-    mode : str, optional
-        'all', 'points' or 'radii' will be scaled
-
-    Returns
-    -------
-    res : (,4)-array
-        scaled fiber
-    """
-
-    fiber = np.array(fiber, copy=True)
-    if mode == 'all':
-        fiber *= scale
-    elif mode == 'points':
-        fiber[:, :3] *= scale
-    elif mode == 'radii':
-        fiber[:, -1] *= scale
-    else:
-        raise ValueError('mode = [all, points, radii]')
-    return fiber
-
-
-def rotate(fiber, rot, offset=None):
-    """
-    Rotates fiber around offset
-
-    Parameters
-    ----------
-    fiber : (,4)-array
-        fiber
-    rot : (3,3)-array_like
-        scale factor
-    offset : 3d-array-array_like, optional
-        offset for rotation center
-
-    Returns
-    -------
-    res : (,4)-array
-        rotated fiber
-    """
-
-    rot = np.array(rot, copy=False)
-    fiber = np.array(fiber, copy=True)
-    if offset is None:
-        fiber[:, :3] = np.dot(rot, fiber[:, :3].T).T
-    else:
-        offset = np.array(offset, copy=False)
-        fiber[:, :3] = np.dot(rot, (fiber[:, :3] - offset).T).T + offset
-    return fiber
-
-
-def translate(fiber, offset):
-    """
-    Translates fiber
-
-    Parameters
-    ----------
-    fiber : (,4)-array
-        fiber
-    offset : 3d-array-array_like
-        offset to translate
-
-    Returns
-    -------
-    res : (,4)-array
-        translated fiber
-    """
-
-    fiber = np.array(fiber, copy=True)
-    offset = np.array(offset, copy=False)
-    fiber[:, :3] += offset
-    return fiber
-
-
-@numba.njit(cache=True)
-def _fiber_segment_aabb_in_aabb(c0, c1, vmin, vmax):
-    c_min = np.array([
-        min(c0[0] - c0[-1], c1[0] - c1[-1]),
-        min(c0[1] - c0[-1], c1[1] - c1[-1]),
-        min(c0[2] - c0[-1], c1[2] - c1[-1])
-    ])
-
-    c_max = np.array([
-        max(c0[0] + c0[-1], c1[0] + c1[-1]),
-        max(c0[1] + c0[-1], c1[1] + c1[-1]),
-        max(c0[2] + c0[-1], c1[2] + c1[-1])
-    ])
-
-    for i in range(3):
-        if c_min[i] > vmax[i] or c_max[i] < vmin[i]:
-            return False
-    return True
-
-
-@numba.njit(cache=True)
-def _fiber_segment_aabb_in_sphere(c0, c1, r, center):
-    c_min = np.array([
-        min(c0[0] - c0[-1], c1[0] - c1[-1]),
-        min(c0[1] - c0[-1], c1[1] - c1[-1]),
-        min(c0[2] - c0[-1], c1[2] - c1[-1])
-    ])
-
-    c_max = np.array([
-        max(c0[0] + c0[-1], c1[0] + c1[-1]),
-        max(c0[1] + c0[-1], c1[1] + c1[-1]),
-        max(c0[2] + c0[-1], c1[2] + c1[-1])
-    ])
-
-    dmin = 0
-    for i in range(3):
[…]
