+++ docs/requirements.txt
@@ -1,4 +1,5 @@
-sphinx
-sphinx-rtd-theme
-sphinx-autodoc-typehints
-numpydoc
+Sphinx==3.4.3
+sphinx-rtd-theme==0.5.1
+sphinx-autodoc-typehints==1.11.1
+sphinx-markdown-tables==0.0.15
+recommonmark==0.7.1
