+++ docs/source/index.rst
@@ -1,15 +1,0 @@
-Welcome to fastPLI
-=======================================
-
-.. autosummary::
-   :toctree: _autosummary
-   :caption: API Reference
-   :template: custom-module-template.rst
-   :recursive:
-
-   fastpli.analysis
-   fastpli.io
-   fastpli.model
-   fastpli.objects
-   fastpli.simulation
-   fastpli.tools
