+++ docs/Makefile
@@ -5,8 +5,8 @@
 # from the environment for the first two.
 SPHINXOPTS    ?=
 SPHINXBUILD   ?= sphinx-build
-SOURCEDIR     = source
-BUILDDIR      = build
+SOURCEDIR     = .
+BUILDDIR      = _build
 
 # Put it first so that "make" without argument is like "make help".
 help:
@@ -17,14 +17,15 @@ help:
 .PHONY: help Makefile
 github:
 	rm -rf gh-pages
-	git clone --branch gh-pages --single-branch https://github.com/3d-pli/fastpli.git gh-pages
 	make html
+	git clone --branch gh-pages --single-branch https://github.com/3d-pli/fastpli.git gh-pages
+	rm -r gh-pages/*
 	cp -r $(BUILDDIR)/html/* gh-pages/
 	touch gh-pages/.nojekyll
 
-
 # Convert .py to .ipynb (don't seem to need to --execute? perhaps sphinx-build does it...), then build html:
 %: Makefile
-	rm -rf build
-	rm -rf source/_autosummary
+	@rm -rf $(BUILDDIR)
+	@rm -rf _autosummary
+	@rm -rf gh-pages 
 	@$(SPHINXBUILD) -M $@ "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
