### 2021-02-15 fdbfd2e
FIX env activation for sphinx docu
