+++ .github/workflows/release.yml
@@ -2,8 +2,7 @@ name: Release-Check
 
 on:
   push:
-    branches:
-      - "release/**"
+    branches: [TestCI]
     tags:
       - "v[0-9].[0-9]+.[0-9]+"
   workflow_dispatch:
@@ -58,6 +57,7 @@ jobs:
       - name: build
         run: |
           make local
+          source env/bin/activate
           make docs
 
       - name: commit
