+++ .github/workflows/release.yml
@@ -1,81 +1,0 @@
-name: Release-Check
-
-on:
-  push:
-    branches: [TestCI]
-    tags:
-      - "v[0-9].[0-9]+.[0-9]+"
-  workflow_dispatch:
-
-jobs:
-  Tests:
-    strategy:
-      matrix:
-        os: [ubuntu-18.04, ubuntu-20.04]
-        config:
-          - { run-opengl: true }
-          - { run-opengl: false }
-
-    runs-on: ${{ matrix.os }}
-
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: dependencies
-        run: |
-          export DEBIAN_FRONTEND=noninteractive
-          sudo apt-get update -qq
-          sudo apt-get install -y gcc g++ clang-format-10 cmake make git
-          sudo apt-get install -y python3-dev python3-venv
-          sudo apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
-
-      - name: openGL
-        run: sudo apt-get install -y freeglut3-dev
-        if: matrix.config.run-it
-
-      - name: run
-        run: |
-          ./CI/run-all.sh
-
-  Sphinx-Doc:
-    needs: Tests
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          persist-credentials: false
-          fetch-depth: 0
-
-      - name: dependencies
-        run: |
-          export DEBIAN_FRONTEND=noninteractive
-          sudo apt-get update -qq
-          sudo apt-get install -y gcc g++ clang-format-10 cmake make git
-          sudo apt-get install -y python3-dev python3-venv
-          sudo apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
-
-      - name: build
-        run: |
-          make local
-          source env/bin/activate
-          make docs
-
-      - name: commit
-        run: |
-          cd docs/gh-pages
-          git config --local user.email "docs@bot"
-          git config --local user.name "docs deployment bot"
-          git status
-          git add -A .
-          git diff-index --quiet HEAD || git commit -m "UPDATE docs"
-
-      - name: push
-        uses: ad-m/github-push-action@master
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          branch: gh-pages
-          directory: docs/gh-pages
-# Zip-Release:
-#   needs: Sphinx-Doc
-#   runs-on: ubuntu-latest
-#   steps:
