+++ docs/index.rst
@@ -1,6 +1,14 @@
+fastPLI API documentation
+=========================
+
 .. autosummary::
    :toctree: _autosummary
    :template: custom-module-template.rst
    :recursive:
 
-   fastpli
+   fastpli.analysis
+   fastpli.io
+   fastpli.model
+   fastpli.objects
+   fastpli.simulation
+   fastpli.tools
