### 2021-02-15 #3042 @whedon
[REVIEW]: fastPLI: A Fiber Architecture Simulation Toolbox for 3D-PLI
**Submitting author:** @fmatuschke (<a href="http://orcid.org/0000-0001-6435-5351">Felix Matuschke</a>)
**Repository:** <a href="https://github.com/3d-pli/fastpli" target ="_blank">https://github.com/3d-pli/fastpli</a>
**Version:** 1.1.0
