### 2021-02-16 f93d4a0
ADD notebook format and check
