+++ CI/run-code-check.sh
@@ -36,6 +36,7 @@ env-CI/bin/python3 -m flake8
 if [ ! $? -eq 0 ]; then EXIT_STATUS=1; fi
 
 # Notebooks
+env-CI/bin/pip3 install -q -r examples/requirements.txt
 find ./examples -iname '*ipynb' | xargs -I {} env-CI/bin/jupyter-nbconvert --clear-output --ClearMetadataPreprocessor.enabled=True --stdout {} | diff {} - &>/dev/null
 if [ ! $? -eq 0 ]; then EXIT_STATUS=1; fi
 
