### 2021-02-16 5065ca6
RM old CI files; RFC examples
