+++ CI/run-examples.sh
@@ -16,13 +16,5 @@ for f in examples/*.py; do
 done
 
 for f in examples/*.ipynb; do
-   echo "converting $f"
-   sed '/^\s*"%matplotlib/d' $f | \
-   sed  's/plt\.show()/# plt\.show()/g' | \
-   env-CI/bin/jupyter-nbconvert --to script --output $f --stdin
-   echo "running $f"
-   env-CI/bin/python3 $f.py
-   rm $f.py
-   echo "done $f"
-   echo ""
+   env-CI/bin/jupyter-nbconvert --execute --to notebook $f
 done
