### 2021-02-16 22b7c2a
RFC cleaned CI scripts
