+++ CI/run-env.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -euo pipefail
-
-python3 -m venv env-CI
