+++ CI/run-all.sh
@@ -2,12 +2,6 @@
 set -euo pipefail
 
 echo "--------------------------------------"
-echo "****************env-CI****************"
-echo "--------------------------------------"
-./CI/run-env.sh
-echo "~~~~~~~~~~~~~~~~~Done~~~~~~~~~~~~~~~~~"
-
-echo "--------------------------------------"
 echo "****************Build*****************"
 echo "--------------------------------------"
 if [ $# -eq 0 ]; then
