### 2021-02-16 632daeb
FIX notebook diff check
