### 2021-02-16 #3042 @vigji
I'll link here below all issues related to this review:

https://github.com/3d-pli/fastpli/issues/1

https://github.com/3d-pli/fastpli/issues/2

https://github.com/3d-pli/fastpli/issues/3

https://github.com/3d-pli/fastpli/issues/4

https://github.com/3d-pli/fastpli/issues/5

https://github.com/3d-pli/fastpli/issues/6
