### 2021-02-17 3cce418
RFC optic_chiasm notebook; ADD python version
