+++ examples/optic_chiasm.py
@@ -1,0 +1,141 @@
+import fastpli.model.sandbox
+import fastpli.model.solver
+import fastpli.simulation
+import fastpli.analysis
+import fastpli.tools
+import fastpli.io
+
+import time
+import matplotlib.pyplot as plt
+import multiprocessing as mp
+import numpy as np
+import scipy.interpolate as si
+
+pool = mp.Pool(2)
+np.random.seed(42)
+
+# two crossing main bundle
+fb_0 = np.array([[-2000, -2000, 0], [-1000, -1000, 0], [0, 0, 0],
+                 [1000, 1000, 0], [2000, 2000, 0]])
+fb_1 = np.array([[-2000, 2000, 0], [-1000, 1000, 0], [0, 0, 0],
+                 [1000, -1000, 0], [2000, -2000, 0]])
+
+# left and right non crossing bundle
+fb_left = np.array([[-2000, -2000, 0], [-1050, -1000, 0], [-300, 0, 0],
+                    [-1050, 1000, 0], [-2000, 2000, 0]])
+fb_right = fb_left.copy()
+fb_right[:, 0] *= -1
+
+# interpolate left and right bundle for smoother
+N = 20
+t = np.linspace(0, 1, fb_left.shape[0])
+t_intp = np.linspace(0, 1, N)
+fb_left = np.array([
+    si.interp1d(t, fb_left[:, 0], 'quadratic')(t_intp),
+    si.interp1d(t, fb_left[:, 1], 'quadratic')(t_intp),
+    si.interp1d(t, fb_left[:, 2], 'quadratic')(t_intp)
+]).T
+fb_right = np.array([
+    si.interp1d(t, fb_right[:, 0], 'quadratic')(t_intp),
+    si.interp1d(t, fb_right[:, 1], 'quadratic')(t_intp),
+    si.interp1d(t, fb_right[:, 2], 'quadratic')(t_intp)
+]).T
+
+fig, ax = plt.subplots(1, 1)
+plt.plot(fb_0[:, 0], fb_0[:, 1])
+plt.plot(fb_1[:, 0], fb_1[:, 1])
+plt.plot(fb_left[:, 0], fb_left[:, 1])
+plt.plot(fb_right[:, 0], fb_right[:, 1])
+ax.axis('equal')
+plt.show()
+
+population = fastpli.model.sandbox.seeds.triangular_circle(800, 200)
+fiber_radii = np.random.uniform(50.0, 100.0, population.shape[0])
+ffb_0 = fastpli.model.sandbox.build.bundle(fb_0, population, fiber_radii)
+ffb_1 = fastpli.model.sandbox.build.bundle(fb_1, population, fiber_radii)
+
+population = fastpli.model.sandbox.seeds.triangular_circle(400, 200)
+fiber_radii = np.random.uniform(50.0, 100.0, population.shape[0])
+ffb_left = fastpli.model.sandbox.build.bundle(fb_left, population, fiber_radii)
+ffb_right = fastpli.model.sandbox.build.bundle(fb_right, population,
+                                               fiber_radii)
+
+# Solving
+solver = fastpli.model.solver.Solver()
+solver.fiber_bundles = [ffb_0, ffb_1, ffb_left, ffb_right]
+solver.obj_min_radius = 150
+solver.obj_mean_length = 150
+solver.omp_num_threads = 2
+
+for i in range(1000):
+    solved = solver.step()
+    if i % (42 // 4) == 0:
+        solver.draw_scene()
+    if solved:
+        break
+solver.draw_scene()
+time.sleep(4.2)
+solver.close_scene()
+
+# Simulation: generate discretised volume
+simpli = fastpli.simulation.Simpli()
+simpli.omp_num_threads = 2
+simpli.voxel_size = 5  # in micro meter
[…]
