### 2021-02-17 272b3ca
ADD orientation analysis to solver notebook
