### 2021-02-18 06b4583
RM old simplui_cpu.py
