+++ src/fastpli/simulation/simpli_cpu.py
@@ -1,459 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Simpli CPU Class
-"""
-
-import warnings
-
-import numpy as np
-
-from .__simpli import __Simpli
-from .__generation import _Generator
-from .__simulation import _Simulator
-
-from . import _mpi
-from .. import analysis
-
-
-class Simpli(__Simpli):
-    """
-    Simpli Class for simulating 3D-PLI images
-    """
-
-    def __init__(self, mpi_comm=None):
-
-        super().__init__()
-
-        # LIBRARIES
-        self.__gen = _Generator()
-        self.__sim = _Simulator()
-        self.mpi = None
-        if mpi_comm:
-            from mpi4py import MPI
-            self.__gen.set_mpi_comm(MPI._addressof(mpi_comm))
-            self.__sim.set_mpi_comm(MPI._addressof(mpi_comm))
-            self.mpi = _mpi._MPI(mpi_comm)
-
-        # freeze class
-        self.__freeze()
-
-    @property
-    def interpolate(self):
-        """ (de)activate interpolation of vectors in simulation: bool """
-        return self._interpolate
-
-    @interpolate.setter
-    def interpolate(self, interpolate):
-        if interpolate not in ('NN', 'Lerp', 'Slerp'):
-            raise ValueError('Only \'NN\', \'Lerp\' or \'Slerp\' are supported')
-        self._interpolate = interpolate
-
-    def generate_tissue(self, only_tissue=False):
-        """ generating discret tissue for simulation """
-
-        self._print('Generate Tissue')
-        self._print(f'Memory needed: ~{np.ceil(self.memory_usage()):.0f} MB')
-
-        self._check_volume_input()
-        self._check_generation_input()
-
-        self.__gen.set_volume(self._dim, self._dim_origin, self._voxel_size)
-        if self._fiber_bundles:
-            self.__gen.set_fiber_bundles(self._fiber_bundles,
-                                         self._fiber_bundles_properties)
-        if self._cells_populations:
-            self.__gen.set_cell_populations(self._cells_populations,
-                                            self._cells_populations_properties)
-        tissue, optical_axis, tissue_properties = self.__gen.run_generation(
-            only_tissue)
-
-        if not np.any(tissue):
-            warnings.warn(
-                'All labels are 0. Usually this means, that the VOI contains no\
-                     fiber_bundles or that the voxel_size is to large.',
-                UserWarning)
-
-        if self.mpi:
-            self.mpi._set_gen_dim(self.__gen.dim_local(), self._dim,
-                                  self.__gen.dim_offset())
-
-        return tissue, optical_axis, tissue_properties
-
-    def _init_pli_setup(self):
-        """ check all pli setup input requirements """
-
-        self._check_simulation_input()
-        self.__sim.set_pli_setup(self._step_size, self._light_intensity,
-                                 self._voxel_size, self._wavelength,
-                                 self._tissue_refrection, self._interpolate,
-                                 self._untilt_sensor_view, self._flip_z_beam,
-                                 self._filter_rotations)
-
-    def _check_generation_input(self):
-        if not self._fiber_bundles and not self._cells_populations:
-            raise ValueError('fiber_bundles and cells_populations are not set')
-        self._check_property_length()
-
-    def run_simulation(self, tissue, optical_axis, tissue_properties, theta,
-                       phi):
-        """
-        running simulation. Input from tissue_generation
-
-        tissue and optical_axis will be passed by reference
-        theta, phi: tilting angle in radiant
-        """
-
-        self._print('Run simulation')
-        self._check_volume_input()
-        self._init_pli_setup()
-
[…]
