+++ Makefile
@@ -111,8 +111,15 @@ docs: ${VENV} clean-docs
 .PHONY: wiki
 wiki: ${VENV}
 	${VENV}/bin/pip3 -q install -r examples/requirements.txt
-	rm -rf examples/wiki
+	rm -rf examples/notebooks
 	find examples/ -iname '*.ipynb' | xargs -P 2 -I {} ${VENV}/bin/jupyter-nbconvert --execute --to markdown --output-dir='examples/notebooks' {}
+	cd examples/notebooks/; \
+	find . -iname '*.md' | xargs -I {} sed -i '2 a ### Note' {} ; \
+	find . -iname '*.md' | xargs -I {} sed -i '3G' {} ; \
+	find . -iname '*.md' | rev | cut -c4- | rev | cut -c3- | xargs -I {} sed -i '4 a > Jupyter notebook version: [.\/examples\/{}.ipynb](https:\/\/github.com\/3d-pli\/fastpli\/blob\/master\/examples\/{}.ipynb)' {}.md ; \
+	find . -iname '*.md' | xargs -I {} sed -i '5G' {} ; \
+	find . -iname '*.md' | cut -c3- | xargs -I {} mv {} tutorial-{}
+
 
 
 .PHONY: docker
