+++ src/fastpli/simulation/__simpli.py
@@ -67,7 +67,7 @@ class __Simpli:
             print(msg)
 
     def get_dict(self):
-        """ Get all member variables which are properties """
+        """ Get all member variables which are properties as dictionary """
         members = dict()
         for key, value in self.__dict__.items():
             if key in ('_fiber_bundles'):
@@ -95,14 +95,14 @@ class __Simpli:
 
     @property
     def dim(self):
-        """ dimension of volume in voxel: (3)-array """
+        """ dimension of volume in voxel: (3)-np.ndarray """
         return self._dim
 
     @dim.setter
     def dim(self, dim):
         dim = np.array(dim)
         if dim.dtype != int:
-            raise TypeError('dim is not np.array(int)')
+            raise TypeError('dim is not np.ndarray(int)')
         if dim.size != 3:
             raise TypeError('dim.size != 3')
         if dim.ndim != 1:
@@ -112,7 +112,7 @@ class __Simpli:
 
     @property
     def dim_origin(self):
-        """ origin of volume in micro: (3)-array """
+        """ origin of volume in micro meter: (3)-np.ndarray """
         return self._dim_origin
 
     @dim_origin.setter
@@ -121,7 +121,7 @@ class __Simpli:
 
     @property
     def voxel_size(self):
-        """ size of voxel in micro: float """
+        """ size length of voxel in micro meter """
         return self._voxel_size
 
     @voxel_size.setter
@@ -143,7 +143,7 @@ class __Simpli:
 
     @property
     def pixel_size(self):
-        """ pixel size of resulting optical image in micro: float """
+        """ side length of pixel of resulting optical image in micro meter """
         return self._pixel_size
 
     @pixel_size.setter
@@ -153,7 +153,15 @@ class __Simpli:
         self._pixel_size = float(pixel_size)
 
     def get_voi(self):
-        """ get volume of interest """
+        """
+        get volume of interest
+
+        Returns
+        -------
+        res: np.ndarray, np.ndarray
+            v_min: [x_min, y_min, z_min]
+            v_max: [x_max, y_max, z_max]
+        """
         if self._voxel_size is None:
             raise ValueError('voxel_size is not set, voi can\'t be calculated')
 
@@ -175,8 +183,12 @@ class __Simpli:
         """
         set volume of interest
 
-        v_min: [x_min, y_min, z_min]
-        v_max: [x_max, y_max, z_max]
+        Parameters
+        ----------
+        v_min: (float,float,float)-array-like
+            [x_min, y_min, z_min]
+        v_max: (float,float,float)-array-like
+            [x_max, y_max, z_max]
         """
 
         v_min = np.array(v_min, dtype=float)
@@ -221,8 +233,12 @@ class __Simpli:
 
     @property
     def tilts(self):
-        """ list of spherical tilting angles [[theta, phi], ...] in radiant:
-        [[float, float], ...]
+        """
+        list of spherical tilting angles [[theta, phi], ...] in radiant:
+
+        Parameters
+        ----------
+        tilts: [[float, float], ...]
         """
         return self._tilts
 
@@ -241,7 +257,13 @@ class __Simpli:
 
     @property
     def noise_model(self):
-        """ applying noise model: function """
+        """
+        noise model to apply on resampled image
+
+        Parameters
+        ----------
+        fun: function
+        """
         return self._noise_model
 
     @noise_model.setter
@@ -255,8 +277,12 @@ class __Simpli:
     @property
     def optical_sigma(self):
         """
-        optical sigma for applying a gaussian convolution to the image after
-        resizing in pixel_size
+        optical_sigma of convolution to applie before resampled image
+
+        Parameters
+        ----------
+        optical_sigma: float
+            in pixel size
         """
         return self._optical_sigma
 
@@ -273,7 +299,14 @@ class __Simpli:
 
[…]
