+++ README.md
@@ -6,46 +6,76 @@ _  __/ / /_/ /_(__  )/ /_ _  ____/_  /____/ /
 /_/    \__,_/ /____/ \__/ /_/     /_____/___/
 -->
 
-# Fiber Architecture Simulation Toolbox for 3D-PLI
+# Fiber Architecture Simulation Toolbox for PLI
 
 ![fastpli-logo](logo.svg)
 
-[fastPLI](https://github.com/3d-pli/fastpli) is an open source toolbox for modeling nerve fibers, simulating them in a [3D-PLI](https://dx.doi.org/10.3389%2Ffninf.2011.00034) microscope and the signal processing developed by the [fiber architecture group](https://www.fz-juelich.de/inm/inm-1/EN/Forschung/Fibre%20Architecture/Fibre%20Architecture_node.html) at the [Forschungszentrum Jülich](https://www.fz-juelich.de) - [INM1](https://www.fz-juelich.de/inm/inm-1/EN/Home/home_node.html).
+The [Fiber Architecture Simulation Toolbox for PLI (fastpli)](https://github.com/3d-pli/fastpli) is a toolbox for [polarized light imaging (PLI)](https://www.fz-juelich.de/inm/inm-1/EN/Forschung/Fibre%20Architecture/Fibre%20Architecture_node.html) with three main purposes:
 
-It consists of three consecutive parts:
+<img align="right" src="https://raw.githubusercontent.com/wiki/3d-pli/fastpli/images/fiber.png" alt="fiber" width="125">
 
-- [Sandbox](https://github.com/3d-pli/fastpli/wiki/Sandbox): building nerve fiber models.
-- [Solver](https://github.com/3d-pli/fastpli/wiki/Solver): solve collisions of nerve fiber models
-- [Simulation](https://github.com/3d-pli/fastpli/wiki/Simulation): 3D-PLI simulations of nerve fiber models
+- [`Sandbox` - designing of nerve fiber models](https://github.com/3d-pli/fastpli/wiki/NerveFiber):
+  The first module allows the user to create different types of nerve fiber bundles and additionally fill them with individual nerve fibers.
 
-In addition, other modules exist to support io and analysis.
+  - [Details](https://github.com/3d-pli/fastpli/wiki/NerveFiber)
+  - [Tutorial](https://github.com/3d-pli/fastpli/wiki//tutorial-sandbox)
 
-## Content
+<img align="right" src="" alt="fiber" width="125">
 
-| module                  | information                                                   |
-| ----------------------- | ------------------------------------------------------------- |
-| `fastpli.analysis`      | analysis of 3D-PLI results                                    |
-| `fastpli.io`            | input/output functions, e.g. to read/save fiber_bundles data  |
-| `fastpli.model.sandbox` | building of simple 3d nerve fiber models                      |
-| `fastpli.model.solver`  | generation of non intersection nerve fiber models             |
-| `fastpli.objects`       | manipulation of fastpli objects (e.g. rotation)               |
-| `fastpli.tools`         | mathematical tools and helper function                        |
-| `fastpli.simulation`    | simulation of fiber models inside a virtual 3D-PLI microscope |
+- [`Solver` - generating collision free models](https://github.com/3d-pli/fastpli/wiki/Solver):
+  The second module takes as input a configuration of nerve fibers and checks them for spatial collisions.
+  Since nerve fibers cannot overlap in reality, one must ensure that the models follow the same rules.
+  The solver module implements a simple algorithm that checks for collisions and, if it finds any, pushes the colliding segments of the fibers slightly apart.
+  This is repeated until all collisions are solved.
 
-## Additional information
[…]
