+++ examples/crossing.py
@@ -1,225 +1,0 @@
-import fastpli.model.sandbox
-import fastpli.model.solver
-import fastpli.simulation
-import fastpli.analysis
-import fastpli.tools
-import fastpli.io
-
-import matplotlib.pyplot as plt
-import matplotlib.animation as animation
-import multiprocessing as mp
-import numpy as np
-import h5py
-import os
-
-pool = mp.Pool(2)
-np.random.seed(42)
-
-FILE_NAME = os.path.abspath(__file__)
-FILE_PATH = os.path.dirname(FILE_NAME)
-FILE_BASE = os.path.basename(FILE_NAME)
-FILE_OUT = os.path.join(FILE_PATH, f'fastpli.example.{FILE_BASE}')
-
-
-def data2image(data):
-    return np.swapaxes(np.flip(data, 1), 0, 1)
-
-
-fiber_bundle_trj_0 = [[-2000, 0, 0], [2000, 0, 0]]
-fiber_bundle_trj_1 = [[0, -2000, 0], [0, 2000, 0]]
-
-population = fastpli.model.sandbox.seeds.triangular_circle(750, 30)
-population = fastpli.model.sandbox.seeds.crop_rectangle([-50, -10000],
-                                                        [50, 10000], population)
-
-fiber_radii = np.random.uniform(10.0, 30.0, population.shape[0])
-fiber_bundle_0 = fastpli.model.sandbox.build.bundle(fiber_bundle_trj_0,
-                                                    population, fiber_radii)
-
-population[:, [0, 1]] = population[:, [1, 0]]
-fiber_radii = np.random.uniform(10.0, 30.0, population.shape[0])
-fiber_bundle_1 = fastpli.model.sandbox.build.bundle(fiber_bundle_trj_1,
-                                                    population, fiber_radii)
-
-solver = fastpli.model.solver.Solver()
-solver.fiber_bundles = [fiber_bundle_0, fiber_bundle_1]
-solver.obj_min_radius = 8 * 20
-solver.obj_mean_length = 4 * 20
-solver.omp_num_threads = 2
-
-# run solver
-print('Begin solving process:')
-solver.draw_scene()
-for i in range(1000):
-    solved = solver.step()
-
-    # calculate current overlap ratio
-    overlap = solver.overlap / solver.num_col_obj if solver.num_col_obj else 0
-    if i % 10 == 0:
-        print(f'step: {i}, {solver.num_obj}/{solver.num_col_obj} ' +
-              f'{round(overlap * 100)}%')
-        solver.draw_scene()
-        # solver.save_ppm(f'solver_{i:03}.ppm')  # save a ppm image
-
-    if i == 20:
-        solver.draw_scene()
-        solver.fiber_bundles = solver.fiber_bundles.apply_to_points(
-            lambda p: p + np.random.uniform(-10, 10, p.shape))
-        solver.fiber_bundles = solver.fiber_bundles.apply_to_radii(
-            lambda r: r * np.random.lognormal(0, 0.1, r.shape))
-        solver.draw_scene()
-
-    if solved:
-        print(f'solved: {i}, {solver.num_obj}/{solver.num_col_obj}')
-        solver.draw_scene()
-        break
-
-fastpli.io.fiber_bundles.save(f'{FILE_OUT}.dat', solver.fiber_bundles, mode='w')
-solver.close_scene()
-
-# Simulation
-print(f'creating file: {FILE_OUT}.h5')
-with h5py.File(f'{FILE_OUT}.h5', 'w') as h5f:
-    # save script
-    h5f['version'] = fastpli.__version__
-    with open(os.path.abspath(__file__), 'r') as f:
-        h5f['parameter/script'] = f.read()
[…]
