### 2021-02-22 3d-pli/fastpli#6 @fmatuschke
I have redesigned my wiki. It took me some time, but I hope it is much clearer now.
I tried to highlight the three basic components of my software. In addition, as you suggested, I have written a tutorial for each part in a Jupyter notebook. These tutorials are available within the wiki as a file converted to Markdown. The same notebook can also be executed by the user from the actual ipynb file in `examples/`.
