### 2021-02-22 3d-pli/fastpli#5 @fmatuschke
I thought about your comment about the objects, and you were right. I wrote the fiber/fiber_bundle and fiber_bundels objects into their own class. All the previous methods are now class functions of the individual objects. In addition, a fiber_bundle has a `layer` property that represents the properties previously set separately.
I tried to explain the functionality within my new wiki and tutorials. Each method is additionally listed in the API documentation.
I hope this makes it easier for users to get started with my software.
