### 2021-02-22 #3042 @vigji
Hey @oliviaguest ! I did most of my comments on the package via issues in the repo, will write a complete report once they will all be addressed - If I got the procedure right!
