### 2021-02-25 3d-pli/fastpli#7 @glyg
git clone should be recursive
In the README, the gi clone line shoud read:
```sh
git clone --recursive https://github.com/3d-pli/fastpli.git
```
