### 2021-02-25 3d-pli/fastpli#8 @glyg
warning during install
Although install proceeded without problems, I had the following warning:
```
WARNING: Built wheel for fastpli is invalid: Metadata 1.2 mandates PEP 440 version, but 'v1.1.0a6.dev1-g2b24015-Release' is not
Failed to build fastpli
Installing collected packages: numpy, llvmlite, scipy, numba, h5py, fastpli
    Running setup.py install for fastpli ... done
  DEPRECATION: fastpli was installed using the legacy 'setup.py install' method, because a wheel could not be built for it. A possible replacement is to fix the wheel build issue reported above. You can find discussion regarding this at https://github.com/pypa/pip/issues/8368.
```

I assume it is a simple fix to not put the git hash in  the version, but I'm not familiar with the system you're using

This is a very minor problem, please feel free to  not fix :)
