### 2021-02-25 3d-pli/fastpli#9 @glyg
ENH: use Ipyvolume or other 3D visualisation
The matplotlib 3D module is not very powerfull, I think data visualization for this library would benefit from native 3D libraries.

In the notebook, ipyvolume is quite simple to setup. In a Qt / desktop context, vispy is well established.
