### 2021-02-25 3d-pli/fastpli#10 @glyg
Use conda-feedstock to distribute the software
This package would benefit from conda-feedstock for binaries cross-compilation and distribution.

I _think_ openmpi and freeglut can be installed as conda packages directly, so the install recipe should not be too complicated to write.
