### 2021-02-25 #3042 @glyg
Hi all

First, congrats @fmatuschke on the software, it is a really nice piece of work.  I opened several issues on fastPLI repo for technical concerns.

Here are aspects more related to the paper itself:

### State of the field:  Do the authors describe how this software compares to other commonly-used packages?

Here I failed to see a review of other similar software or efforts. The cited literature focuses on the authors' team. Although I understand the imaging modality is specific, is it not used in other settings? Simulation of IRM data is alluded to, maybe this adjacent field has more references? As is it is hard to situate this work in the general field of brain/neuron imagery.
A more scholarly (minor) question: I know 2 photons / second harmonics generation methods are also used to image bi-refringent tissues. Could fastPLI be adapted to modeling this modality? Are there comparable modeling efforts for  these techniques?

### Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

Some revision of the writing should be made. @oliviaguest should I provide a detailed list of typos and such?

### References: Is the list of references complete, and is everything cited appropriately that should be cited

As stated in the first section, a review of the field or references to contributions from outside the authors' team (unless I missed something) would be welcome.

### Opened issues:

I opened 4 issues

The first 2 require minor efforts [#7](https://github.com/3d-pli/fastpli/issues/7), [#8](https://github.com/3d-pli/fastpli/issues/8) and are cosmetic.

The next 2 ([#9](https://github.com/3d-pli/fastpli/issues/9), [#10](https://github.com/3d-pli/fastpli/issues/10) ) would improve the software distribution and visualization, but I don't consider they need to be solved prior to publication.

In summary, I think this paper should be published  pending the minor revisions in paper.md outlined above.
