### 2021-02-26 fd71cbf
FIX crop_circle, wrong radii
