### 2021-03-02 3d-pli/fastpli#7 @fmatuschke
Thanks, that is a good suggestion. I added it to the last commit.
