### 2021-03-02 3d-pli/fastpli#8 @fmatuschke
That's correct. For the purpose of traceability I automatically generate the version number from the last tag version name. However, if a new commit has been used without a version tag, the additional number of commits after the last tag is added (here 1) and the git hash will be added additionally. If someone decides to use a commit, e.g. in another development branch, this warning also reminds them. However, I should make it more obvious than the this.

Clearly in the future I will only push/pull request new releases of the code. For the purpose of the current review process I will push also little changes, so all reviewers can see the changes. Therefore, you can ignore this warning at this point.
