### 2021-03-02 3d-pli/fastpli#9 @fmatuschke
That's a good suggestion. I will look into this in the future.
I added matplotlib only for the use of the examples, since "everybody" knows it. As you can see, it is not a requirement for my package.
For the collision solver process I added an implementation in OpenGL as described in the wiki <https://github.com/3d-pli/fastpli/wiki/Solver> which is fast enough (compared to the solving process) to help the user to see what's going on. However, it is not interactively. The libraries you mentioned can be a nice improvement for the future.
