### 2021-03-02 3d-pli/fastpli#4 @fmatuschke
The publishing process of the API documentation seems to work fine. I can of course be further improved. At this point I think however I can close this issue. Thanks again for the suggestion.
