### 2021-03-02 #3042 @fmatuschke
@glyg thanks for the compliment. Since this is my first official attempt at releasing software, it means a lot.
All the issues you opened in the software repository were commented by me. They are nice suggestions for improvements.
As for the missing discussion of other software, I definitely failed to address that. I will add this as soon as possible and let you all know.
