+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-CMAKE_MINIMUM_REQUIRED(VERSION 3.9)
+CMAKE_MINIMUM_REQUIRED(VERSION 3.12)
 
 PROJECT( fastpli LANGUAGES CXX)
 SET(CMAKE_COLOR_MAKEFILE ON)
