+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-CMAKE_MINIMUM_REQUIRED(VERSION 3.12)
+CMAKE_MINIMUM_REQUIRED(VERSION 3.9)
 
 PROJECT( fastpli LANGUAGES CXX)
 SET(CMAKE_COLOR_MAKEFILE ON)
@@ -77,7 +77,9 @@ ENDIF()
 # Libraries
 ##########################################################################
 # Python
-FIND_PACKAGE(Python3 COMPONENTS Interpreter Development REQUIRED)
+FIND_PACKAGE(PythonInterp 3.6 REQUIRED)
+SET(PYTHON3_INTP_VERSION "${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}.${PYTHON_VERSION_PATCH}")
+FIND_PACKAGE(PythonLibs ${PYTHON3_INTP_VERSION} REQUIRED)
 
 ADD_SUBDIRECTORY(lib/pybind11)
 IF (CMAKE_CXX_COMPILER_ID STREQUAL "Clang")
