+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'fastPLI: A Fiber Architecture Simulation Toolbox for 3D-PLI'
+title: "fastPLI: A Fiber Architecture Simulation Toolbox for 3D-PLI"
 tags:
   - 3D-PLI
   - Python
@@ -18,7 +18,7 @@ authors:
 affiliations:
   - name: Institute of Neuroscience and Medicine (INM-1), Forschungszentrum Jülich GmbH, 52425, Jülich, Germany
     index: 1
-  - name: Cécile and Oskar Vogt Institute for Brain Research, University Hospital Düsseldorf, University of Düsseldorf, 40204, Düsseldorf, Germany. 
+  - name: Cécile and Oskar Vogt Institute for Brain Research, University Hospital Düsseldorf, University of Düsseldorf, 40204, Düsseldorf, Germany.
     index: 2
 date: 18. December 2020
 bibliography: paper.bib
@@ -32,36 +32,36 @@ The physical effect behind 3D-PLI is the optical property of the nerve fibers ca
 Due to this intrinsic birefringence, it is possible to use polarized light, pass it through a thin brain section and observe the change of the polarization state of light.
 This change is directly related to the 3D orientation of the fibers and also provides strong contrasts between fibers and other tissue components.
 
-To understand the influence of the underlying fiber structure, simulations are an essential tool. They allow testing different hypotheses, knowing the ground truth. It has already been shown that simulations with scattered light within tissue sections require models with irregularities to mimic the behavior of the scattered light. This knowledge can now be used to understand structures such as fiber crossings that have been difficult to interpret in 3D PLI [@Menzel2020].
+To understand the influence of the underlying fiber structure, simulations are an essential tool. They allow testing different hypotheses, knowing the ground truth. It has already been shown that simulations with scattered light within tissue sections require models with irregularities to mimic the behavior of the scattered light. This knowledge can now be used to understand structures such as fiber crossings that have been difficult to interpret in 3D-PLI [@Menzel2020].
 
 In addition, the generated nerve fiber models can be used in other imaging simulation techniques such as diffusion magnetic resonance imaging.
 
 # Summary
 
-*fastPLI* is an open source toolbox based on Python and C++ for modeling myelinated axons, i.e. nerve fibers and simulating the results of measurement of fiber orientations with a polarization microscope using 3D-PLI.
+_fastPLI_ is an open source toolbox based on Python and C++ for modeling myelinated axons, i.e. nerve fibers and simulating the results of measurement of fiber orientations with a polarization microscope using 3D-PLI.
 
-The *fastPLI* package includes the following modules:
+The _fastPLI_ package includes the following modules:
 
-1) **Fiber Modelling Modules:**
-    A detailed 3D modelling of nerve fibers at the micrometer level is essential as input for the measurement simulation.
-    In order to recreate biological tissue as a model, it is important that the nerve fibers do not spatially overlap.
-    We have decided to implement a solver module that takes any configuration of fiber bundles as input and converts it over several iterations into an collision-free configuration.
-    In order to generate collision free fiber arrangements, a dedicated algorithm to prohibit such overlaps has been developed [@Matuschke2019].
+1. **Fiber Modelling Modules:**
+   A detailed 3D modelling of nerve fibers at the micrometer level is essential as input for the measurement simulation.
+   In order to recreate biological tissue as a model, it is important that the nerve fibers do not spatially overlap.
[…]
