+++ paper/paper.bib
@@ -106,3 +106,91 @@
   booktitle = {Proceedings of the Second Workshop on the {LLVM} Compiler Infrastructure in {HPC} - {LLVM} {\textquotesingle}15},
   keywords = {software}
 }
+
+@inproceedings{Chapelle2015,
+title = "Generation of non-overlapping fiber architecture",
+keywords = "Random periodic packing, Curved fibers, Fiber orientation",
+author = "Lucie Chapelle and M. L{\'e}vesque and Povl Br{\o}ndsted and Foldschack, {M. R.} and Yukihiro Kusano",
+year = "2015",
+language = "English",
+booktitle = "Proceedings of the 20th International Conference on Composite Materials",
+publisher = "ICCM20 Secretariat",
+note = "20th International Conference on Composite Materials (ICCM20) ; Conference date: 19-07-2015 Through 24-07-2015",
+}
+
+@article{Altendorf2011,
+  doi = {10.1103/physreve.83.041804},
+  url = {https://doi.org/10.1103/physreve.83.041804},
+  year = {2011},
+  month = apr,
+  publisher = {American Physical Society ({APS})},
+  volume = {83},
+  number = {4},
+  author = {Hellen Altendorf and Dominique Jeulin},
+  title = {Random-walk-based stochastic modeling of three-dimensional fiber systems},
+  journal = {Physical Review E}
+}
+
+@article{Mingasson2017,
+  doi = {10.3389/fninf.2017.00005},
+  url = {https://doi.org/10.3389/fninf.2017.00005},
+  year = {2017},
+  month = jan,
+  publisher = {Frontiers Media {SA}},
+  volume = {11},
+  author = {Tom Mingasson and Tanguy Duval and Nikola Stikov and Julien Cohen-Adad},
+  title = {{AxonPacking}: An Open-Source Software to Simulate Arrangements of Axons in White Matter},
+  journal = {Frontiers in Neuroinformatics}
+}
+
+@article{Ginsburger2019,
+  doi = {10.1016/j.neuroimage.2019.02.055},
+  url = {https://doi.org/10.1016/j.neuroimage.2019.02.055},
+  year = {2019},
+  month = jun,
+  publisher = {Elsevier {BV}},
+  volume = {193},
+  pages = {10--24},
+  author = {K{\'{e}}vin Ginsburger and Felix Matuschke and Fabrice Poupon and Jean-Fran{\c{c}}ois Mangin and Markus Axer and Cyril Poupon},
+  title = {{MEDUSA}: A {GPU}-based tool to create realistic phantoms of the brain microstructure using tiny spheres},
+  journal = {{NeuroImage}}
+}
+
+@article{RamellaRoman2005,
+  doi = {10.1364/opex.13.004420},
+  url = {https://doi.org/10.1364/opex.13.004420},
+  year = {2005},
+  publisher = {The Optical Society},
+  volume = {13},
+  number = {12},
+  pages = {4420},
+  author = {Jessica C. Ramella-Roman and Scott A. Prahl and Steve L. Jacques},
+  title = {Three Monte Carlo programs of polarized light transport into scattering media: part I},
+  journal = {Optics Express}
+}
+
+@article{vanTurnhout2009,
+  doi = {10.1117/1.3241986},
+  url = {https://doi.org/10.1117/1.3241986},
+  year = {2009},
+  publisher = {{SPIE}-Intl Soc Optical Eng},
+  volume = {14},
+  number = {5},
+  pages = {054018},
+  author = {Mark C. van Turnhout and Sander Kranenbarg and Johan L. van Leeuwen},
[…]
