+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 # Statement of need
 
 3D-PLI is a microscopic neuroimaging technique used to study the nerve fiber architecture in unstained histological brain sections at the micrometer scale [@Axer2011].
-It allows to give contrast to fibers and fiber tracts and, ultimately, reconstruct 3D nerve fiber orientations.
+It allows contrasting fibers and fiber tracts and ultimately reconstructing 3D nerve fiber orientations.
 The physical effect behind 3D-PLI is the optical property of the nerve fibers called birefringence.
 Due to this intrinsic birefringence, it is possible to use polarized light, pass it through a thin brain section and observe the change of the polarization state of light.
 This change is directly related to the 3D orientation of the fibers and also provides strong contrasts between fibers and other tissue components.
@@ -44,7 +44,7 @@ While these algorithms are specialized in their field, _fastPLI_ also provides a
 Here, the focus is on the previously developed algorithm [@Matuschke2019], which provides a fast method to generate collision-free results for white matter structures in the brain.
 
 Different types of simulation for polarized light are for example described in [RamellaRoman2005; @vanTurnhout2009; @Jiang2020].
-However, to our knowledge, none of these techniques have been used to simulate the effects of polarized light on nerve fibers, with the exception of `simPLI` [@Dohmen2015], which is included in this toolbox.
+However, to our knowledge, none of these techniques have been used to simulate the effects of polarized light on nerve fibers, except `simPLI` [@Dohmen2015], which is included in this toolbox.
 
 # Summary
 
@@ -55,7 +55,7 @@ The _fastPLI_ package includes the following modules:
 1. **Fiber Modelling Modules:**
    A detailed 3D modelling of nerve fibers at the micrometer level is essential as input for the measurement simulation.
    In order to recreate biological tissue as a model, it is important that the nerve fibers do not spatially overlap.
-   We have decided to implement a solver module that takes any configuration of fiber bundles as input and converts it over several iterations into an collision-free configuration.
+   We have decided to implement a solver module that takes any configuration of fiber bundles as input and converts it over several iterations into a collision-free configuration.
    In order to generate collision free fiber arrangements, a dedicated algorithm to prohibit such overlaps has been developed [@Matuschke2019].
 
 2. **Simulation Module:**
@@ -64,7 +64,7 @@ The _fastPLI_ package includes the following modules:
    By this means the brain section can be scanned from oblique views which adds important information to unambiguously analyze the 3D fiber orientation.
 
 3. **Analysis Module:**
-   The resulting simulated measurements (i.e., image stacks of a section acquired at different polarizing filter rotation angles and, optionally, at different oblique views) can be processed in a similar way as the real, experimental 3D-PLI [@Axer2011; @Schmitz2018].
+   The resulting simulated measurements (i.e., image stacks of a section acquired at different polarizing filter rotation angles and, optionally, at different oblique views) can be processed similarly to the real, experimental 3D-PLI [@Axer2011; @Schmitz2018].
 
 All computationally intensive calculations are optimized either with _numba_ on the Python side or with multithreading _C++_ algorithms, which can be accessed via _pybind11_ inside the Python package [@Lam2015;@pybind11].
 Additionally, the simulation module supports the Message Passing Interface (MPI) to facilitate the simulation of very large volumes on multiple computer nodes.
@@ -110,10 +110,10 @@ An example of the solving process is shown in fig. \ref{generation} a).
 The red colored segments indicate that a collision with this segment is detected.
 With further iterations, the number of collisions decreases.
 At the end, a collision-free configurations results.
[…]
