### 2021-03-15 3d-pli/fastpli#11 @RealPolitiX
CMake error at installation
I got an error trying to compile with cmake on Ubuntu 18.04 (Python==3.7.3)
```
CMake Error at CMakeLists.txt:80 (FIND_PACKAGE):
  By not providing "FindPython3.cmake" in CMAKE_MODULE_PATH this project has
  asked CMake to find a package configuration file provided by "Python3", but
  CMake did not find one.

  Could not find a package configuration file provided by "Python3" with any
  of the following names:

    Python3Config.cmake
    python3-config.cmake

  Add the installation prefix of "Python3" to CMAKE_PREFIX_PATH or set
  "Python3_DIR" to a directory containing one of the above files.  If
  "Python3" provides a separate development package or SDK, be sure it has
  been installed.

-- Configuring incomplete, errors occurred!
See also "/mnt/pcshare/users/rpx/gitpod/fastpli/build/CMakeFiles/CMakeOutput.log".
Makefile:36: recipe for target 'build/Makefile' failed
make: *** [build/Makefile] Error 1
```
