### 2021-03-15 #3042 @RealPolitiX
The paper from @fmatuschke is well-written apart from a few grammar errors. The references are thorough. I think due to the novelty of the measurement technique used for reconstruction of neuronal fibers, it's not used elsewhere in general. You can probably cite some of the references from [OpenPolScope](https://openpolscope.org/pages/Literature.htm), since it's related and discuss your unique context.

As for the software, I opened [issue #11](https://github.com/3d-pli/fastpli/issues/11). The wiki looks nicely written.
