### 2021-03-15 3d-pli/fastpli#11 @fmatuschke
Hi, I noticed this error as well. CMake depricated the `FindPythonInterp` and `FindPythonLibs` files to the newer method `FindPython3` in the version `3.12`.
The CMake version in Ubuntu 18.04 LTS is `3.10.2` to old.
However since I want at least to support Ubuntu 18.04 LTS and Python 3.6 I switched to the old methods in 89696587ae08c5f421bb9094395b29d839e63ccb.
Interestingly this error was not catched by the CI.
