### 2021-03-15 3d-pli/fastpli#11 @fmatuschke
I will leave this Issue open until I found a nice solution which supports both CMake versions. Additionally I noticed in the past, there can be some Issues with multiple installed Python versions. That was also the reason, why I switched to the new method.
