### 2021-03-15 #3042 @fmatuschke
I added the missing section about the software compares mentioned by @glyg.
Additionally I fixed some typos I found.

Thank you @RealPolitiX for opening https://github.com/3d-pli/fastpli/issues/11. I fixed it.

Just a small summary for @all.
I have addressed all the issues so far. The main changes since the opening of the review are a complete rework of my Wiki and API documentation.
