+++ CMakeLists.txt
@@ -77,10 +77,15 @@ ENDIF()
 # Libraries
 ##########################################################################
 # Python
-FIND_PACKAGE(PythonInterp 3.6 REQUIRED)
-SET(PYTHON3_INTP_VERSION "${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}.${PYTHON_VERSION_PATCH}")
-FIND_PACKAGE(PythonLibs ${PYTHON3_INTP_VERSION} REQUIRED)
+IF(${CMAKE_VERSION} VERSION_LESS "3.12.0") 
+   FIND_PACKAGE(PythonInterp 3.6 REQUIRED)
+   SET(Python3_VERSION "${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}.${PYTHON_VERSION_PATCH}")
+   FIND_PACKAGE(PythonLibs ${PYTHON3_INTP_VERSION} REQUIRED)
+ELSE()
+   FIND_PACKAGE(Python3 COMPONENTS Interpreter Development REQUIRED)
+ENDIF()
 
+# Pybind11
 ADD_SUBDIRECTORY(lib/pybind11)
 IF (CMAKE_CXX_COMPILER_ID STREQUAL "Clang")
    SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fsized-deallocation" )
