### 2021-03-16 3d-pli/fastpli#5 @vigji
Thanks for addressing this! Yes, the new wiki addresses everything, I'll close this.
