### 2021-03-22 8080504
ADD python3 macro to Makefile; add flake8 ignore patterns
