+++ CMakeLists.txt
@@ -82,7 +82,7 @@ IF(${CMAKE_VERSION} VERSION_LESS "3.12.0")
    SET(Python3_VERSION "${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}.${PYTHON_VERSION_PATCH}")
    FIND_PACKAGE(PythonLibs ${PYTHON3_INTP_VERSION} REQUIRED)
 ELSE()
-   FIND_PACKAGE(Python3 COMPONENTS Interpreter Development REQUIRED)
+   FIND_PACKAGE(Python3 3.6 COMPONENTS Interpreter Development REQUIRED)
 ENDIF()
 
 # Pybind11
