+++ CMakeLists.txt
@@ -77,13 +77,19 @@ ENDIF()
 # Libraries
 ##########################################################################
 # Python
-IF(${CMAKE_VERSION} VERSION_LESS "3.12.0") 
-   FIND_PACKAGE(PythonInterp 3.6 REQUIRED)
+
+# SET(PYTHON3_VERSION "" CACHE STRING "")
+
+MESSAGE("Python3_VERSION: ${Python3_VERSION}")
+
+IF(${CMAKE_VERSION} VERSION_LESS "3.12.0")
+   FIND_PACKAGE(PythonInterp EXACT ${Python3_VERSION} REQUIRED)
    SET(Python3_VERSION "${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}.${PYTHON_VERSION_PATCH}")
-   FIND_PACKAGE(PythonLibs ${PYTHON3_INTP_VERSION} REQUIRED)
+   FIND_PACKAGE(PythonLibs EXACT ${Python3_VERSION} REQUIRED)
 ELSE()
-   FIND_PACKAGE(Python3 3.6 COMPONENTS Interpreter Development REQUIRED)
+   FIND_PACKAGE(Python3 EXACT ${Python3_VERSION} COMPONENTS Interpreter Development REQUIRED)
 ENDIF()
+LIST(APPEND FASTPLI_INFO_LIBRARIES "Python-v${Python3_VERSION}")
 
 # Pybind11
 ADD_SUBDIRECTORY(lib/pybind11)
@@ -91,7 +97,7 @@ IF (CMAKE_CXX_COMPILER_ID STREQUAL "Clang")
    SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fsized-deallocation" )
 ENDIF()
 INCLUDE_DIRECTORIES(${Python3_INCLUDE_DIRS})
-LIST(APPEND FASTPLI_INFO_LIBRARIES "Python-v${Python3_VERSION}")
+LIST(APPEND FASTPLI_INFO_LIBRARIES "Pybind-v${PYBIND11_PYTHON_VERSION}")
 
 # OpenMP
 FIND_PACKAGE( OpenMP )
