### 2021-03-31 7677a79
FIX type conversion error in epa and simpli for integer measurments
