+++ src/fastpli/analysis/epa.py
@@ -21,9 +21,10 @@ def epa(data):
     """
 
     data = np.array(data, copy=False)
-
     n = data.shape[-1]
-    rho_2 = 2 * np.linspace(0, np.pi, n, False, dtype=data.dtype)
+
+    dtype = np.float32 if data.itemsize <= 4 else np.float64
+    rho_2 = 2 * np.linspace(0, np.pi, n, False, dtype=dtype)
 
     a0 = np.sum(data, -1) / n
     a1 = 2 * np.sum(data * np.sin(rho_2), -1) / n
@@ -53,9 +54,10 @@ def direction(data):
     """
 
     data = np.array(data, copy=False)
-
     n = data.shape[-1]
-    rho_2 = 2 * np.linspace(0, np.pi, n, False, dtype=data.dtype)
+
+    dtype = np.float32 if data.itemsize <= 4 else np.float64
+    rho_2 = 2 * np.linspace(0, np.pi, n, False, dtype=dtype)
 
     a1 = 2 * np.sum(data * np.sin(rho_2), -1) / n
     b1 = 2 * np.sum(data * np.cos(rho_2), -1) / n
