### 2021-04-16 #3042 @vigji
I am very sorry, yes, I was missing the notifications!

I went again through the repository. For the review, I opened 6 issues (https://github.com/3d-pli/fastpli/issues/1, https://github.com/3d-pli/fastpli/issues/2, https://github.com/3d-pli/fastpli/issues/3, https://github.com/3d-pli/fastpli/issues/4, https://github.com/3d-pli/fastpli/issues/5, https://github.com/3d-pli/fastpli/issues/6) concerning installation, automatic testing, clarity of the documentation, and internal organisation and naming consistency in the package. All of them have been thoroughly addressed by the author. Kudos to the new wiki, it is extensive and very readable now! I could successfully install the package and ran the tests.

I checked the last version of the manuscript and it looks good to me. I don't have the expertise in the subject of interest to suggest other related packages and publications that should be discussed in the paper,  but the literature now discussed in the context of fibres models seems good to me!

All my issues have been addressed and I have no further requirements for the paper to be published. Nice job @fmatuschke !
