+++ docs/conf.py
@@ -2,7 +2,7 @@
 #
 # This file only contains a selection of the most common options. For a full
 # list see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
+# https://www.sphinx-doc.org/en/main/usage/configuration.html
 
 # -- Path setup --------------------------------------------------------------
 
@@ -53,7 +53,6 @@ add_module_names = False  # Remove namespaces from class/method signatures
 napoleon_google_docstring = False
 napoleon_use_param = False
 napoleon_use_ivar = True
-
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
