+++ .gitlab/.gitlab-ci.yml
@@ -1,33 +1,32 @@
 image: ubuntu:20.04
 
 stages:
-   - test
+  - test
 
 test:release:
-   stage: test
-   only: 
-      - development
-      - master
-      - TestCI
-   tags:
-      - docker
-   except:
-      variables:
-         - $CI_COMMIT_MESSAGE =~ /WIP/
-   script:
-      - export DEBIAN_FRONTEND=noninteractive
-      - apt-get update -qq
-      - apt-get install -y gcc g++ clang-format-10 cmake make git
-      - apt-get install -y python3-dev python3-venv
-      - apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
-      - apt-get install -y freeglut3-dev
-      - CI/run-all.sh
-
+  stage: test
+  only:
+    - development
+    - main
+    - TestCI
+  tags:
+    - docker
+  except:
+    variables:
+      - $CI_COMMIT_MESSAGE =~ /WIP/
+  script:
+    - export DEBIAN_FRONTEND=noninteractive
+    - apt-get update -qq
+    - apt-get install -y gcc g++ clang-format-10 cmake make git
+    - apt-get install -y python3-dev python3-venv
+    - apt-get install -y libopenmpi-dev libhdf5-openmpi-dev h5utils
+    - apt-get install -y freeglut3-dev
+    - CI/run-all.sh
 # test:no_opengl:
 #    stage: test
-#    only: 
+#    only:
 #       - development
-#       - master
+#       - main
 #       - TestCI
 #    tags:
 #       - docker
