+++ paper/paper.bib
@@ -38,17 +38,16 @@
   journal = {Physical Review X}
 }
 
-@article{Krause2016,
-  doi = {10.17815/jlsrf-2-121},
-  url = {https://doi.org/10.17815/jlsrf-2-121},
-  year = {2016},
-  month = mar,
-  publisher = {Forschungszentrum Julich,  Zentralbibliothek},
-  volume = {2},
-  author = {Dorian Krause and Philipp Th\"{o}rnig},
-  title = {{JURECA}: General-purpose supercomputer at J\"{u}lich Supercomputing Centre},
-  journal = {Journal of large-scale research facilities {JLSRF}}
-}
+@article {jureca,
+  author = {{J\"{u}lich Supercomputing Centre}},
+  title  = {{JURECA: Modular supercomputer at J\"{u}lich Supercomputing Centre}},
+  journal = {Journal of large-scale research facilities},
+  number  = {A132},
+  volume  = {4},
+  doi     = {10.17815/jlsrf-4-121-1},
+  url     = {http://dx.doi.org/10.17815/jlsrf-4-121-1},
+  year    = {2018}
+} 
 
 @article{Schmitz2018,
   doi = {10.3389/fnana.2018.00075},
