### 2021-04-26 #3042 @fmatuschke
@all I fixed hopefully the last typo and one reference in the paper.
@oliviaguest I published the software in zenodo with the DOI: [10.5281/zenodo.4720075](https://zenodo.org/record/4720075)
