### 2021-04-30 #3042 @Kevin-Mattheus-Moerman
@fmatuschke I have proofread your paper and have some very minor comments which I hope you can address:

- [ ] Please check `It allows contrasting fibers and fiber tracts and ultimately reconstructing 3D nerve fiber orientations`, some words seem missing. Should this be something like `It provides image contrast for fibers and fiber tracts, and ultimately enables reconstruction of 3D nerve fiber orientations`?
- [ ] Please introduce the acronym dMRI for diffusion magnetic resonance imaging, when the terms are first used/introduced.
- [ ] Consider changing `Different types of simulation...` to `Different types of simulations...`

Once you have considered the above please call `@whedon generate pdf` to update the paper. I will then review the changes and process acceptance when ready.

One more important note relates to the Zenodo archive
- [ ] Please ensure that the meta-data for the Zenodo archive matches that of the paper in terms of the author list as well as the title. Currently the archive title read: `Fiber Architecture Simulation Toolbox for 3D-PLI` but this should match the paper and be changed to `fastPLI: A Fiber Architecture Simulation Toolbox for3D-PLI`. You can manually change this on Zenodo.

The version tag for the Zenodo archive is `1.1.0`. I will therefore assume we should update the version tag here also. I will do this shortly below :point_down:.

Thanks.
