### 2021-05-03 #3042 @fmatuschke
Dear @Kevin-Mattheus-Moerman,
I change all three of your comments. Thanks for your help.
Regarding the version number. Yes, it is now officially 1.1. Thanks to the reviewers I changed quite a bit, so I had to increase the minor version.
I have also corrected the title in the Zenodo archive.
