### 2021-05-03 #3042 @Kevin-Mattheus-Moerman
Congratulations @fmatuschke this work is now published in JOSS.
Thank you @oliviaguest for editing this submission!

Thanks also to @vigji, @glyg, @RealPolitiX for your help reviewing this work!!!
